### 2024-04-12 #6608 @editorialbot
[PRE REVIEW]: Variational Solvers for Irreversible Evolutionary Systems
**Submitting author:** @kumiori (<a href="http://orcid.org/0000-0002-3019-602X">Andrés A. León Baldelli</a>)
**Repository:** https://github.com/kumiori/irrevolutions
**Branch with paper.md** (empty if default branch): andres-paper
**Version:** 0.1.0
