### 2024-04-23 7f03fa7
benchmaking umut, some cleaning and viz
