### 2024-04-25 #6608 @majensen
@mkhorton, would you be willing to review [this work](#6608) for [JOSS](https://joss.theoj.org)? It may interest you in light of your work at LBNL. Really appreciate your consideration -- @majensen
