+++ playground/IDRIS-CAMPAING-HYDRA/scripts/1d-traction-first-order-legacy-algorithm.py
@@ -8,6 +8,7 @@ from pathlib import Path
 from typing import Optional
 
 from libidris.core import (elastic_energy_density, damage_energy_density, stress)
+from libidris.core import a
 
 
 import dolfinx
@@ -112,8 +113,12 @@ def run_computation(parameters, storage=None):
     Lx = parameters.get("geometry").get("Lx")
 
     # Define the state
-    u_zero = Function(V_u, name="InelasticDisplacement")
+    # u_zero = Function(V_u, name="InelasticDisplacement")
     zero_u = Function(V_u, name="BoundaryUnknown")
+    zero_u.interpolate(lambda x: np.zeros_like(x[0]))
+    
+    tilde_u = Function(V_u, name="BoundaryDatum")
+    tilde_u.interpolate(lambda x: np.ones_like(x[0]))
 
     # Measures
     dx = ufl.Measure("dx", domain=mesh)
@@ -126,22 +131,23 @@ def run_computation(parameters, storage=None):
     alpha_ub.interpolate(lambda x: np.ones_like(x[0]))
 
     eps_t = dolfinx.fem.Constant(mesh, np.array(1., dtype=PETSc.ScalarType))
-    u_zero.interpolate(lambda x: eps_t/2. * (2*x[0] - Lx))
+    # u_zero.interpolate(lambda x: eps_t/2. * (2*x[0] - Lx))
     
-    for f in [u, zero_u, u_zero, alpha_lb, alpha_ub]:
+    for f in [u, zero_u, tilde_u, alpha_lb, alpha_ub]:
         f.vector.ghostUpdate(
             addv=PETSc.InsertMode.INSERT, mode=PETSc.ScatterMode.FORWARD
         )
 
-    bcs_u = [dirichletbc(u_zero, dofs_u_right), 
-             dirichletbc(u_zero, dofs_u_left)]
+    bcs_u = [dirichletbc(np.array(0, dtype=PETSc.ScalarType), dofs_u_left, V_u), 
+             dirichletbc(tilde_u, dofs_u_right)]
 
     # bcs_u = []
     bcs_alpha = []
     
     bcs = {"bcs_u": bcs_u, "bcs_alpha": bcs_alpha}
     
-    total_energy = (elastic_energy_density(state, parameters, u_zero) + damage_energy_density(state, parameters)) * dx
+    total_energy = (elastic_energy_density(state, parameters) \
+        + damage_energy_density(state, parameters)) * dx
     
     load_par = parameters["loading"]
     loads = np.linspace(load_par["min"], load_par["max"], load_par["steps"])
@@ -158,17 +164,19 @@ def run_computation(parameters, storage=None):
 
     logging.basicConfig(level=logging.INFO)
 
+    time_series = []
+    alpha_values = []
+    displacement_tip_values = []
+
 
     for i_t, t in enumerate(loads):
 
-        eps_t.value = t
-        
-        u_zero.interpolate(lambda x: eps_t/2. * (2*x[0] - Lx))
-        u_zero.vector.ghostUpdate(
-            addv=PETSc.InsertMode.INSERT, mode=PETSc.ScatterMode.FORWARD
-        )
+        # eps_t.value = t
 
-        # update the lower bound
+        tilde_u.interpolate(lambda x: t * np.ones_like(x[0]))
+        tilde_u.vector.ghostUpdate(addv=PETSc.InsertMode.INSERT,
+                            mode=PETSc.ScatterMode.FORWARD)
+        
         alpha.vector.copy(alpha_lb.vector)
         alpha_lb.vector.ghostUpdate(
             addv=PETSc.InsertMode.INSERT, mode=PETSc.ScatterMode.FORWARD
[…]
