+++ playground/IDRIS-CAMPAING-HYDRA/scripts/1d-traction-first-order.py
@@ -1,0 +1,425 @@
+
+#!/usr/bin/env python3
+import hashlib
+import json
+import logging
+import os
+import sys
+from pathlib import Path
+from typing import Optional
+
+import dolfinx
+import dolfinx.mesh
+import dolfinx.plot
+import numpy as np
+import pandas as pd
+import petsc4py
+import pyvista
+import ufl
+import yaml
+from dolfinx.common import list_timings
+from dolfinx.fem import (Constant, Function, assemble_scalar, dirichletbc,
+                         form, locate_dofs_geometrical, set_bc)
+from dolfinx.fem.petsc import assemble_vector, set_bc
+from dolfinx.io import XDMFFile
+from irrevolutions.algorithms.am import HybridSolver
+from irrevolutions.algorithms.so import BifurcationSolver, StabilitySolver
+from irrevolutions.solvers import SNESSolver
+from irrevolutions.solvers.function import vec_to_functions
+from irrevolutions.test.test_1d import _AlternateMinimisation1D as am1d
+from irrevolutions.utils import (ColorPrint, ResultsStorage, Visualization,
+                                 _logger, _write_history_data, history_data,
+                                 norm_H1, norm_L2)
+from irrevolutions.utils.plots import (plot_AMit_load, plot_energies,
+                                       plot_force_displacement)
+from irrevolutions.utils.viz import (plot_mesh, plot_profile, plot_scalar,
+                                     plot_vector)
+from mpi4py import MPI
+from petsc4py import PETSc
+from pyvista.utilities import xvfb
+
+from irrevolutions.utils.viz import _plot_bif_spectrum_profiles
+
+petsc4py.init(sys.argv)
+comm = MPI.COMM_WORLD
+
+# Mesh on node model_rank and then distribute
+model_rank = 0
+
+
+def run_computation(parameters, storage=None):
+    Lx = parameters["geometry"]["Lx"]
+    _nameExp = parameters["geometry"]["geom_type"]
+    parameters["model"]["ell"]
+
+    # Get geometry model
+    parameters["geometry"]["geom_type"]
+    _N = int(parameters["geometry"]["N"])
+
+    mesh = dolfinx.mesh.create_unit_interval(MPI.COMM_WORLD, _N)
+    outdir = os.path.join(os.path.dirname(__file__), "output")
+
+    if storage is None:
+        prefix = os.path.join(outdir, f"thin-film-at2")
+    else:
+        prefix = storage
+
+    if comm.rank == 0:
+        Path(prefix).mkdir(parents=True, exist_ok=True)
+
+    signature = hashlib.md5(str(parameters).encode("utf-8")).hexdigest()
+
+    if comm.rank == 0:
+        with open(f"{prefix}/parameters.yaml", "w") as file:
+            yaml.dump(parameters, file)
+
+        with open(f"{prefix}/signature.md5", "w") as f:
+            f.write(signature)
+
+    with XDMFFile(
+        comm, f"{prefix}/{_nameExp}.xdmf", "w", encoding=XDMFFile.Encoding.HDF5
+    ) as file:
+        file.write_mesh(mesh)
+
+    # Functional Setting
+    element_u = ufl.FiniteElement("Lagrange", mesh.ufl_cell(), degree=1)
+    element_alpha = ufl.FiniteElement("Lagrange", mesh.ufl_cell(), degree=1)
+
+    V_u = dolfinx.fem.FunctionSpace(mesh, element_u)
+    V_alpha = dolfinx.fem.FunctionSpace(mesh, element_alpha)
+
+    u = dolfinx.fem.Function(V_u, name="Displacement")
[…]
