### 2024-05-15 2dc4194
IDRIS first-order new hybrid
