+++ playground/IDRIS-CAMPAING-HYDRA/scripts/1d-film-first-order-new-hybrid-natural.py
@@ -7,60 +7,9 @@ import sys
 from pathlib import Path
 from typing import Optional
 
-from libidris.core import (elastic_energy_density, damage_energy_density, stress)
+from libidris.core import (elastic_energy_density_film, damage_energy_density, stress)
 from libidris.core import a
-
-"""
-Numerical Experiment: Fracture in a Film on Substrate
-
-Description:
--------------
-This numerical experiment simulates fracture phenomena in a thin film adhered to a substrate (e.g. silicon on glass). 
-The goal is to analyse the initiation and emergence of patterns of cracks under tensile loading and to understand the mechanical response of the material system.
-
-Experiment:
------------------------
-- Objective: To study the crack nucleation and pattern emergence. 
-- Solver: First order only, legacy alternate mininiation algorithm.
-- Hypothesis: The sound (homogeneous) solution is expected lose stability, or not. Emergence of crack pattern through the film and not into the substrate.
-- Validation: The results will be compared against experimental data from literature, or not.
-- Question: Does the system miss the bifurcation point?
-
-Additional Information:
-------------------------
-- Mesh Details: A refined mesh with a minimum element size of 0.01 mm is used around the crack tip.
-- Solver Settings: The simulation uses a quasi-static solver with a convergence criterion of 1e-6.
-- Data Analysis: Stress distribution plots and crack path visualization will be performed using post-processing tools.
-
-Material Model:
-----------------
-- Film Material: brittle, 2d asymptotic multilayer system, cf. DOI 10.1007/s10659-015-9528-3
-    eg. Silicon (Young's modulus = 130 GPa, Poisson's ratio = 0.28, fracture toughness = 0.9 MPa√m)
-- Substrate Material: elastic
-    Glass (Young's modulus = 70 GPa, Poisson's ratio = 0.22)
-- Interface Properties: Perfect adhesion assumed between the film and substrate.
-
-Boundary Conditions:
----------------------
-- Geometry: 1d Film (dimensions normalised): 
-    Tp model: 100 mm x 10 mm x 0.1 mm, Substrate dimensions: 100 mm x 10 mm x 1 mm
-- Fixed Boundaries: Natural 
-    The bottom surface of the substrate is fixed in all directions.
-- Symmetry Boundaries: 
-    No symmetry boundary conditions applied.
-
-Initial Conditions:
----------------------
-- Natural state: Sound, unloaded 
-
-Loading:
----------
-- Type of Loading: Uniaxial tensile load, u_zero
-- Magnitude: The tensile load monotonically increases.
-- Loading Rate: The load is applied quasi-statically, no dynamic effects, no inherent time-scales.
-
-
-"""
+from libidris.core import setup_output_directory, save_parameters, create_function_spaces_1d, initialize_functions
 
 import dolfinx
 import dolfinx.mesh
@@ -76,10 +25,8 @@ from dolfinx.fem import (Constant, Function, assemble_scalar, dirichletbc,
                          form, locate_dofs_geometrical, set_bc)
 from dolfinx.fem.petsc import assemble_vector, set_bc
 from dolfinx.io import XDMFFile
-from irrevolutions.algorithms.am import AlternateMinimisation
-from irrevolutions.solvers import SNESSolver
+from irrevolutions.algorithms.am import HybridSolver
 from irrevolutions.solvers.function import vec_to_functions
[…]
