### 2024-05-15 46106ed
practice scripts develop, adding description and fine tuning
