### 2024-05-16 4b1e269
refactoring postprocessing and output in 1d film
