+++ playground/IDRIS-CAMPAING-HYDRA/scripts/1d-film-second-order-stab-kick.py
@@ -35,6 +35,7 @@ from irrevolutions.utils import (ColorPrint, ResultsStorage, Visualization,
                                  norm_H1, norm_L2)
 from irrevolutions.utils.plots import (plot_AMit_load, plot_energies,
                                        plot_force_displacement)
+import matplotlib
 from irrevolutions.utils.viz import (plot_mesh, plot_profile, plot_scalar,
                                      plot_vector)
 from mpi4py import MPI
@@ -182,92 +183,144 @@ def run_computation(parameters, storage=None):
 
         stable = stability.solve(alpha_lb, eig0=z0, inertia=inertia)
 
+        fracture_energy, elastic_energy = postprocess(parameters,
+                                                        _nameExp,
+                                                        prefix,
+                                                        v,
+                                                        β,
+                                                        state,
+                                                        u_zero,
+                                                        dx,
+                                                        bifurcation,
+                                                        stability,
+                                                        i_t)
 
-        import matplotlib
+        dump_output(_nameExp, prefix, history_data, u, alpha, equilibrium, bifurcation, stability, t, inertia, fracture_energy, elastic_energy)
+
+    print(pd.DataFrame(history_data).drop(columns=["cone_data", "eigs_cone", "stable"]))
+    return history_data, {}, state
+
+def dump_output(_nameExp, 
+                    prefix, 
+                    history_data, 
+                    u, 
+                    alpha, 
+                    equilibrium, 
+                    bifurcation, 
+                    stability, 
+                    t, 
+                    fracture_energy, 
+                    elastic_energy):
+    
+    with dolfinx.common.Timer(f"~Output and Storage") as timer:
+        with XDMFFile(
+                    comm, f"{prefix}/{_nameExp}.xdmf", "a", encoding=XDMFFile.Encoding.HDF5
+                ) as file:
+            file.write_function(u, t)
+            file.write_function(alpha, t)
+
+        if comm.rank == 0:
+            a_file = open(f"{prefix}/time_data.json", "w")
+            json.dump(history_data, a_file)
+            a_file.close()
+
+        _write_history_data(
+                equilibrium = equilibrium,
+                bifurcation = bifurcation,
+                stability = stability,
+                history_data = history_data,
+                t=t,
+                inertia = bifurcation.get_inertia(),
+                stable = np.nan,
+                energies = [elastic_energy, fracture_energy],
+            )
+
+def postprocess(parameters, _nameExp, prefix, β, v, state, u_zero, dx, bifurcation, stability, i_t, matplotlib):
+    with dolfinx.common.Timer(f"~Postprocessing and Vis") as timer:
+    
         fig_state, ax1 = matplotlib.pyplot.subplots()
         
-        with dolfinx.common.Timer(f"~Postprocessing and Vis") as timer:
-            if comm.rank == 0:
-                plot_energies(history_data, file=f"{prefix}/{_nameExp}_energies.pdf")
-                plot_AMit_load(history_data, file=f"{prefix}/{_nameExp}_it_load.pdf")
-                # plot_force_displacement(
-                #     history_data, file=f"{prefix}/{_nameExp}_stress-load.pdf"
-                # )
+        if comm.rank == 0:
+            plot_energies(history_data, file=f"{prefix}/{_nameExp}_energies.pdf")
+            plot_AMit_load(history_data, file=f"{prefix}/{_nameExp}_it_load.pdf")
+                    # plot_force_displacement(
+                    #     history_data, file=f"{prefix}/{_nameExp}_stress-load.pdf"
+                    # )
 
-            xvfb.start_xvfb(wait=0.05)
-            pyvista.OFF_SCREEN = True
+        xvfb.start_xvfb(wait=0.05)
+        pyvista.OFF_SCREEN = True
 
-            plotter = pyvista.Plotter(
[…]
