### 2024-05-16 28524c6
kick works with variational stepper in 1d film
