+++ playground/IDRIS-CAMPAING-HYDRA/scripts/1d-film-second-order-stab-kick.py
@@ -27,7 +27,7 @@ from dolfinx.fem.petsc import assemble_vector, set_bc
 from dolfinx.io import XDMFFile
 from irrevolutions.algorithms.am import HybridSolver
 from irrevolutions.algorithms.so import BifurcationSolver, StabilitySolver
-from irrevolutions.algorithms.ls import StabilityStepper
+from irrevolutions.algorithms.ls import StabilityStepper, LineSearch
 
 from irrevolutions.solvers.function import vec_to_functions
 from irrevolutions.test.test_1d import _AlternateMinimisation1D as am1d
@@ -43,6 +43,7 @@ from mpi4py import MPI
 from petsc4py import PETSc
 from pyvista.utilities import xvfb
 import random
+import matplotlib.pyplot as plt
 
 from irrevolutions.utils.viz import _plot_bif_spectrum_profiles
 
@@ -54,13 +55,6 @@ logger = logging.getLogger(__name__)
 
 # Mesh on node model_rank and then distribute
 model_rank = 0
-class DummyStabilitySolver:
-    def solve(self, t):
-        if t == 0:
-            return True
-        else:
-            return random.choice([True, False])
-
 
 def run_computation(parameters, storage=None):
     Lx = parameters["geometry"]["Lx"]
@@ -160,13 +154,16 @@ def run_computation(parameters, storage=None):
     stability = StabilitySolver(
         total_energy, state, bcs, cone_parameters=parameters.get("stability")
     )
+    
+    linesearch = LineSearch(
+        total_energy,
+        state,
+        linesearch_parameters=parameters.get("stability").get("linesearch"),
+    )
+    
 
-    logging.basicConfig(level=logging.INFO)
-
-
-    dummy_stability = DummyStabilitySolver()
-
-
+    arclength = []
+    
     while True:
         try:
             i_t, t = next(iterator)
@@ -206,69 +203,137 @@ def run_computation(parameters, storage=None):
         _logger.info(f"Stability of state at load {t:.2f}: {stable}")
         
         if not stable:
-            __import__('pdb').set_trace()
             
             iterator.pause_time()
             _logger.info(f"Time paused at {t:.2f}")
+            vec_to_functions(stability.solution["xt"], [v, β])
+            perturbation = {"v": v, "beta": β}
+            interval = linesearch.get_unilateral_interval(state, perturbation)
 
-            # y_t = perturb_state(y_t)
-            # logger.info(f"State perturbed at load {t:.2f}: {y_t}")
+            order = 4
+            h_opt, energies_1d, p, _ = linesearch.search(
+                state, perturbation, interval, m=order
+            )
+            arclength.append((i_t, t, h_opt))
+            
+            with dolfinx.common.Timer(f"~Visualisation") as timer:
+
+                _logger.critical(f" *> State is unstable: {not stable}")
+                _logger.critical(f"line search interval is {interval}")
+                _logger.critical(f"perturbation energies: {energies_1d}")
+                _logger.critical(f"hopt: {h_opt}")
+                _logger.critical(f"lambda_t: {stability.solution['lambda_t']}")
+
+                h_steps = np.linspace(interval[0], interval[1], order + 1)
+                fig, axes = plt.subplots(1, 1)
+                plt.scatter(h_steps, energies_1d)
+                plt.scatter(
+                    h_opt,
+                    0,
+                    c="k",
+                    s=40,
+                    marker="|",
+                    label=f"$h^*={h_opt:.2f}$")
+                plt.scatter(h_opt, p(h_opt), c="k", s=40, alpha=0.5)
+                xs = np.linspace(interval[0], interval[1], 30)
+                axes.plot(xs, p(xs), label="Energy slice along perturbation")
+                axes.set_xlabel("h")
+                axes.set_ylabel("$E_h - E_0$")
[…]
