+++ playground/IDRIS-CAMPAING-HYDRA/scripts/2d-film-first-order-new-algorithm.py
@@ -1,0 +1,425 @@
+#!/usr/bin/env python3
+import hashlib
+import json
+import logging
+import os
+import sys
+from pathlib import Path
+from typing import Optional
+
+from libidris.core import (elastic_energy_density_film, damage_energy_density, stress)
+from libidris.core import a
+from libidris.core import setup_output_directory, save_parameters, create_function_spaces_2d, initialize_functions
+
+import dolfinx
+import dolfinx.mesh
+import dolfinx.plot
+import numpy as np
+import pandas as pd
+import petsc4py
+import pyvista
+import ufl
+import yaml
+from dolfinx.common import list_timings
+from dolfinx.fem import (Constant, Function, assemble_scalar, dirichletbc,
+                         form, locate_dofs_geometrical, set_bc)
+from dolfinx.fem.petsc import assemble_vector, set_bc
+from dolfinx.io import XDMFFile, gmshio
+from irrevolutions.algorithms.am import HybridSolver
+from irrevolutions.solvers.function import vec_to_functions
+from irrevolutions.utils import (ColorPrint, ResultsStorage, Visualization,
+                                 _logger, _write_history_data, history_data,
+                                 norm_H1, norm_L2)
+from irrevolutions.utils.plots import (plot_AMit_load, plot_energies,
+                                       plot_force_displacement)
+from irrevolutions.utils.viz import (plot_mesh, plot_profile, plot_scalar,
+                                     plot_vector)
+from irrevolutions.models import BrittleMembraneOverElasticFoundation as ThinFilm
+from irrevolutions.meshes.primitives import mesh_circle_gmshapi
+
+from mpi4py import MPI
+from petsc4py import PETSc
+from pyvista.utilities import xvfb
+
+from irrevolutions.utils.viz import _plot_bif_spectrum_profiles
+
+petsc4py.init(sys.argv)
+comm = MPI.COMM_WORLD
+
+# Mesh on node model_rank and then distribute
+model_rank = 0
+
+
+class ThinFilmWithImposedDisplacement(ThinFilm):
+    
+    def __init__(self, model_parameters={}, u_0=0):
+        """
+        Initialie material parameters.
+        * Sound material:
+            - E_0: sound Young modulus
+            - K: elastic foundation modulus
+            - nu_0: sound plasticity ratio
+            - u_0: substrate deformation
+            - sig_d_0: sound damage yield stress
+            - ell: internal length
+            - k_res: residual stiffness
+        """
+        # Initialize elastic parameters
+        super().__init__(model_parameters)
+        if model_parameters:
+            self.model_parameters.update(model_parameters)
+        
+        # Initialize the damage parameters
+        self.w1 = self.model_parameters["w1"]
+        self.ell = self.model_parameters["ell"]
+        self.ell_e = self.model_parameters["ell_e"]
+        self.k_res = self.model_parameters["k_res"]
+        self.u_0 = u_0
+
+    def elastic_energy_density(self, state):
+        """
+        Returns the elastic energy density from the state.
+        """
+        # Parameters
+        alpha = state["alpha"]
+        u = state["u"]
+        u_0 = self.u_0 
+        return self.elastic_energy_density_strain(
+            self.eps(u), alpha) + self.elastic_foundation_density(u - u_0)
+
+
+def run_computation(parameters, storage=None):
+    _nameExp = parameters["geometry"]["geom_type"]
+    R = parameters["geometry"]["R"]
[…]
