+++ playground/IDRIS-CAMPAING-HYDRA/scripts/2d-film-second-order-bifurcation.py
@@ -26,6 +26,9 @@ from dolfinx.fem import (Constant, Function, assemble_scalar, dirichletbc,
 from dolfinx.fem.petsc import assemble_vector, set_bc
 from dolfinx.io import XDMFFile, gmshio
 from irrevolutions.algorithms.am import HybridSolver
+from irrevolutions.algorithms.so import BifurcationSolver, StabilitySolver
+from irrevolutions.algorithms.ls import StabilityStepper, LineSearch
+
 from irrevolutions.solvers.function import vec_to_functions
 from irrevolutions.utils import (ColorPrint, ResultsStorage, Visualization,
                                  _logger, _write_history_data, history_data,
@@ -178,7 +181,29 @@ def run_computation(parameters, storage=None):
         solver_parameters=parameters.get("solvers"),
     )
 
-    for i_t, t in enumerate(loads):
+    bifurcation = BifurcationSolver(
+        total_energy, state, bcs, bifurcation_parameters=parameters.get("stability")
+    )
+
+    stability = StabilitySolver(
+        total_energy, state, bcs, cone_parameters=parameters.get("stability")
+    )
+    
+    linesearch = LineSearch(
+        total_energy,
+        state,
+        linesearch_parameters=parameters.get("stability").get("linesearch"),
+    )
+    
+    iterator = StabilityStepper(loads)
+    arclength = []
+    
+    while True:
+        try:
+            i_t, t = next(iterator)
+        except StopIteration:
+            break
+        
         eps_t.value = t
         u_zero.interpolate(lambda x: radial_field(x) * eps_t)
         u_zero.vector.ghostUpdate(
@@ -193,68 +218,49 @@ def run_computation(parameters, storage=None):
         _logger.critical(f"-- Solving for t = {t:3.2f} --")
         with dolfinx.common.Timer(f"~First Order: Equilibrium") as timer:
             equilibrium.solve(alpha_lb)
+
+        is_unique = bifurcation.solve(alpha_lb)
+        is_elastic = not bifurcation._is_critical(alpha_lb)
+        inertia = bifurcation.get_inertia()
+        
+        z0 = (
+            bifurcation._spectrum[0]["xk"]
+            if bifurcation._spectrum and "xk" in bifurcation._spectrum[0]
+            else None
+        )
+        ColorPrint.print_bold(f"Evolution is unique: {is_unique}")
+        ColorPrint.print_bold(f"State's inertia: {inertia}")
+        # stable = stability.solve(alpha_lb, eig0=z0, inertia=inertia)
+        equilibrium.log()
+        bifurcation.log()
         
         with dolfinx.common.Timer(f"~Postprocessing and Vis") as timer:
+
+
+            # magnitude_expr = dolfinx.fem.Expression(ufl.dot(u, u), V_alpha.element.interpolation_points())
+            magnitude_expr = dolfinx.fem.Expression(u[0], V_alpha.element.interpolation_points())
+            magnitude = dolfinx.fem.Function(V_alpha)
+            magnitude.interpolate(magnitude_expr)
+
             if comm.rank == 0:
                 plot_energies(history_data, file=f"{prefix}/{_nameExp}_energies.pdf")
                 plot_AMit_load(history_data, file=f"{prefix}/{_nameExp}_it_load.pdf")
-                # plot_force_displacement(
-                #     history_data, file=f"{prefix}/{_nameExp}_stress-load.pdf"
-                # )
 
             xvfb.start_xvfb(wait=0.05)
             pyvista.OFF_SCREEN = True
 
-            # plotter = pyvista.Plotter(
-            #     title="Profiles",
-            #     window_size=[800, 600],
-            #     shape=(1, 1),
[…]
