### 2024-05-17 4bd5f4d
iterator test extends to minus initial step
