### 2024-05-24 bbc20f8
script updates
