+++ playground/IDRIS-CAMPAING-HYDRA/scripts/1d-traction-first-order-new-hybrid.py
@@ -333,7 +333,7 @@ if __name__ == "__main__":
         model="at1")
     
     # Run computation
-    _storage = f"../output/1d-traction-first-new-hybrid/MPI-{MPI.COMM_WORLD.Get_size()}/{signature}"
+    _storage = f"../output/1d-traction-first-new-hybrid/MPI-{MPI.COMM_WORLD.Get_size()}/{signature[0:6]}"
     visualization = Visualization(_storage)
 
     with dolfinx.common.Timer(f"~Computation Experiment") as timer:
