### 2024-06-16 #6608 @majensen
@jhale would you have time and inclination to review [this work](#6608) for [JOSS](https://joss.theoj.org)? Your help would be much appreciated - thanks for your consideration --@majensen
