### 2024-06-16 #6608 @majensen
Hello @pefarrell - are you available to review [this work](#6608) for [JOSS](https://joss.theoj.org)? Thanks very much for your consideration- --@majensen
