### 2024-06-16 #6608 @jorgensd
> @jorgensd would you have time to review [this work](#6608) for [JOSS](https://joss.theoj.org)? I would appreciate your expertise. Thanks for considering --@majensen

Yes, I can act as reviewer.
