### 2024-06-28 #6897 @jhale
GNU recommends/instructs that LGPLv3 files should be named `COPYING` and `COPYING.LESSER`.
