### 2024-07-10 kumiori/irrevolutions#52 @jorgensd
JOSS references
Comments regarding https://github.com/openjournals/joss-reviews/issues/6897
**References**
- Missing DOI: 10.1016/S0022-5096(99)00028-9
- Missing DOI: 10.1145/1089014.1089019
- DOI rendering error on: León Baldelli, A. A., & Maurini, C. (2021). Numerical bifurcation and stability analysis of153
variational gradient-damage models for phase-field fracture. Journal of the Mechanics154
and Physics of Solids, 152, 104424. https://doi.org/https://doi.org/10.1016/j.jmps.2021.155
104424156
- Missing DOI: 10.5802/crmeca.156
- Missing DOI: 10.1016/j.jmps.2011.03.010
