### 2024-07-10 kumiori/irrevolutions#53 @jorgensd
JOSS manuscript comments
Only minor comments wrt to the manuscript from https://github.com/openjournals/joss-reviews/issues/6897

**Summary**
- I would remove the last sentence of the summary

**Statement of need**
- Do you have a reference for the statement
  >   The current literature in computational fracture mechanics predominantly focuses on unilateral first-order criteria, systematically neglecting the exploration of higher-order information for critical points.

**Three solvers**
- I would rename this to **Functionality**
 - As a non-specialist, it would beneficial to me introduce the problem you would like to solve ([1]) prior to introducing the solvers.
 - Additionally, $V_0$ is not defined (I guess a suitable Sobolev space).
 - I guess it would be good to refer to some papers that have examples of a formulation similar to eq  [1]. For instance the aforementioned paper (https://doi.org/10.1016/j.jmps.2021.104424, section 2.4?). Maybe there are some other references for non-experts that you know?

**Software**
The interface is well described in this section, only question from reading the paper and not testing the implementation is:
- How does `state # A dictionary of fields describing the system` link to `A list of upper bounds (upper and lower) for the state`? As a dictionary is not ordered, and one has multiple states, it is not clear which bound is applied to which state.
- `numerical_parameters` is not described by example, would be useful to see an example in the paper.

**Verification**
I guess you should formulate $z=(v, \beta)$ as $z$ is not explicitly defined in the definition of the Rayleigh ratio.

To me it would be worthwhile rehashing what the exact solutions from the appendix of Pham, K., Marigo, J.-J., & Maurini, C. (2011) is as I am not a domain expert. It would make figure 2-5 easier to interpret. Especially figure 3-5 contains a lot of information that I do not how to easily digest as a non-specialist.

** State of the field: Do the authors describe how this software compares to other commonly-used packages?**
Some references of other software solving similar problems with be useful (maybe linked to my comment in **Statement of need**.
