### 2024-07-10 kumiori/irrevolutions#54 @jorgensd
Contributors and authorship
Looking at the contributor graph: https://github.com/kumiori/irrevolutions/graphs/contributors
there seems to be several additional authors that have contributed significant pieces to the code that are not listed as authors (Andrés A León Baldelli and Pierluigi Cesana).

Could you specify:
- Pierluigi's contributions to the software, as he only has two commits to the repository.
- How the contributions of:  Ayoub RHOUSNI,  SarahTACHET relate to software (are these contributions only examples of usage?)

Ref: https://github.com/openjournals/joss-reviews/issues/6897
