### 2024-07-10 kumiori/irrevolutions#55 @jorgensd
Add installation instructions of irrevolutions to README
Ref: https://github.com/openjournals/joss-reviews/issues/6897

Currently, the installation instructions in readme are out of date, as it

> docker run --rm -ti -v "$(pwd)":/home/numerix -w /home/numerix kumiori3/numerix:latest
ships DOLFINx v0.6.0, while the code uses up to date syntax for DOLFINx (v0.7.2). The CI uses: `kumiori3/numerix:stable-amd64`.

The instructions does not state that one has to `pip install` the software, could be added as something like the following:

_To install the software, run DOLFINx through the aforementioned docker container and install the `irrevolutions`-package with `python3 -m pip install .` from the root of this repository_

I would suggest adding a pin to the DOLFINx version to the dependencies in https://github.com/kumiori/irrevolutions/blob/main/pyproject.toml#L19, i.e. `"fenics-dolfinx==0.7.*"` to ensure that users are running a compatible version of DOLFINx.
