### 2024-07-10 kumiori/irrevolutions#56 @jorgensd
Tests and test structure
Ref: https://github.com/openjournals/joss-reviews/issues/6897

I cannot get several of the tests to run locally, as i get:

```bash
root@e29591a5575e:/home/numerix/irrevolutions/src/irrevolutions/test# pytest -xvs .
======================================= test session starts ========================================
platform linux -- Python 3.10.12, pytest-7.4.2, pluggy-1.3.0 -- /usr/bin/python3
cachedir: .pytest_cache
rootdir: /home/numerix/irrevolutions
configfile: pytest.ini
plugins: anyio-4.3.0, xdist-3.3.1
collecting ... GITHUB_REF environment variable not found.
GITHUB_SHA environment variable not found.
collecting 1 item                                                                                  Info    : Meshing 1D...
Info    : [  0%] Meshing curve 0 (Line)
Info    : [ 30%] Meshing curve 1 (Line)
Info    : [ 50%] Meshing curve 2 (Line)
Info    : [ 80%] Meshing curve 3 (Line)
Info    : Done meshing 1D (Wall 0.000182015s, CPU 0.000167s)
Info    : Meshing 2D...
Info    : Meshing surface 1 (Plane, Frontal-Delaunay)
Info    : Done meshing 2D (Wall 0.00397538s, CPU 0.003975s)
Info    : 360 nodes 722 elements
  0 SNES Function norm 0.000000000000e+00
  0 SNES Function norm 0.000000000000e+00
ALTMIN - Iterations:   1,            Error: 0.0000e+00, alpha_H1,            alpha_max: 0.0000e+00
Saving the eigenvectors for the following eigenvalues
[]
State is elastic: True
State's inertia: (0, 0, 720)
Path is unique: True

[MPI_MAX] Summary of timings                                                |  reps  wall avg  wall tot
-------------------------------------------------------------------------------------------------------
Build dofmap data                                                           |     4  0.000158  0.000633
Build sparsity                                                              |    14  0.000037  0.000517
Compute connectivity 1-2                                                    |     1  0.000004  0.000004
Compute dof reordering map                                                  |     4  0.000004  0.000018
Compute entities of dim = 1                                                 |     1  0.000249  0.000249
Compute local part of mesh dual graph                                       |     1  0.000173  0.000173
Compute local-to-local map                                                  |     1  0.000018  0.000018
Compute-local-to-global links for global/local adjacency list               |     1  0.000004  0.000004
Distribute row-wise data (scalable)                                         |     1  0.000019  0.000019
GPS: create_level_structure                                                 |     4  0.000006  0.000023
Gibbs-Poole-Stockmeyer ordering                                             |     1  0.000058  0.000058
Init dofmap from element dofmap                                             |     4  0.000014  0.000057
SparsityPattern::finalize                                                   |     5  0.000233  0.001167
Topology: create                                                            |     1  0.000258  0.000258
Topology: determine shared index ownership                                  |     1  0.000071  0.000071
Topology: determine vertex ownership groups (owned, undetermined, unowned)  |     1  0.000013  0.000013
~First Order: AltMin-Damage solver                                          |     1  0.000144  0.000144
~First Order: AltMin-Elastic solver                                         |     1  0.000303  0.000303
~Second Order: Bifurcation                                                  |     1  0.943233  0.943233
  0 SNES Function norm 3.909186116588e-01
  1 SNES Function norm 3.271666527291e-01
  2 SNES Function norm 1.996627348698e-01
  3 SNES Function norm 3.341981445294e-16
  0 SNES Function norm 0.000000000000e+00
ALTMIN - Iterations:   1,            Error: 0.0000e+00, alpha_H1,            alpha_max: 0.0000e+00
Saving the eigenvectors for the following eigenvalues
[]
State is elastic: True
State's inertia: (0, 0, 720)
Path is unique: True

[MPI_MAX] Summary of timings                                                |  reps  wall avg  wall tot
-------------------------------------------------------------------------------------------------------
Build dofmap data                                                           |     4  0.000158  0.000633
Build sparsity                                                              |    26  0.000035  0.000911
Compute connectivity 1-2                                                    |     1  0.000004  0.000004
Compute dof reordering map                                                  |     4  0.000004  0.000018
Compute entities of dim = 1                                                 |     1  0.000249  0.000249
Compute local part of mesh dual graph                                       |     1  0.000173  0.000173
Compute local-to-local map                                                  |     1  0.000018  0.000018
Compute-local-to-global links for global/local adjacency list               |     1  0.000004  0.000004
Distribute row-wise data (scalable)                                         |     1  0.000019  0.000019
GPS: create_level_structure                                                 |     4  0.000006  0.000023
Gibbs-Poole-Stockmeyer ordering                                             |     1  0.000058  0.000058
Init dofmap from element dofmap                                             |     4  0.000014  0.000057
SparsityPattern::finalize                                                   |     8  0.000268  0.002142
Topology: create                                                            |     1  0.000258  0.000258
Topology: determine shared index ownership                                  |     1  0.000071  0.000071
Topology: determine vertex ownership groups (owned, undetermined, unowned)  |     1  0.000013  0.000013
~First Order: AltMin-Damage solver                                          |     2  0.000115  0.000229
~First Order: AltMin-Elastic solver                                         |     2  0.001830  0.003660
~Second Order: Bifurcation                                                  |     2  0.488039  0.976078
  0 SNES Function norm 3.909186116588e-01
  1 SNES Function norm 3.271666527291e-01
  2 SNES Function norm 1.996627348698e-01
  3 SNES Function norm 3.984105306175e-16
  0 SNES Function norm 0.000000000000e+00
ALTMIN - Iterations:   1,            Error: 0.0000e+00, alpha_H1,            alpha_max: 0.0000e+00
Saving the eigenvectors for the following eigenvalues
[]
State is elastic: True
State's inertia: (0, 0, 720)
Path is unique: True

[MPI_MAX] Summary of timings                                                |  reps  wall avg  wall tot
-------------------------------------------------------------------------------------------------------
Build dofmap data                                                           |     4  0.000158  0.000633
Build sparsity                                                              |    38  0.000036  0.001362
Compute connectivity 1-2                                                    |     1  0.000004  0.000004
Compute dof reordering map                                                  |     4  0.000004  0.000018
Compute entities of dim = 1                                                 |     1  0.000249  0.000249
Compute local part of mesh dual graph                                       |     1  0.000173  0.000173
Compute local-to-local map                                                  |     1  0.000018  0.000018
Compute-local-to-global links for global/local adjacency list               |     1  0.000004  0.000004
Distribute row-wise data (scalable)                                         |     1  0.000019  0.000019
GPS: create_level_structure                                                 |     4  0.000006  0.000023
Gibbs-Poole-Stockmeyer ordering                                             |     1  0.000058  0.000058
Init dofmap from element dofmap                                             |     4  0.000014  0.000057
SparsityPattern::finalize                                                   |    11  0.000279  0.003069
Topology: create                                                            |     1  0.000258  0.000258
Topology: determine shared index ownership                                  |     1  0.000071  0.000071
Topology: determine vertex ownership groups (owned, undetermined, unowned)  |     1  0.000013  0.000013
~First Order: AltMin-Damage solver                                          |     3  0.000106  0.000317
~First Order: AltMin-Elastic solver                                         |     3  0.002202  0.006607
~Second Order: Bifurcation                                                  |     3  0.336151  1.008454
  0 SNES Function norm 3.909186116588e-01
  1 SNES Function norm 3.271666527291e-01
  2 SNES Function norm 1.996627348698e-01
  3 SNES Function norm 5.624130914496e-16
  0 SNES Function norm 0.000000000000e+00
ALTMIN - Iterations:   1,            Error: 0.0000e+00, alpha_H1,            alpha_max: 0.0000e+00
Saving the eigenvectors for the following eigenvalues
[]
State is elastic: True
State's inertia: (0, 0, 720)
Path is unique: True

[MPI_MAX] Summary of timings                                                |  reps  wall avg  wall tot
-------------------------------------------------------------------------------------------------------
Build dofmap data                                                           |     4  0.000158  0.000633
Build sparsity                                                              |    50  0.000036  0.001784
Compute connectivity 1-2                                                    |     1  0.000004  0.000004
Compute dof reordering map                                                  |     4  0.000004  0.000018
Compute entities of dim = 1                                                 |     1  0.000249  0.000249
Compute local part of mesh dual graph                                       |     1  0.000173  0.000173
Compute local-to-local map                                                  |     1  0.000018  0.000018
Compute-local-to-global links for global/local adjacency list               |     1  0.000004  0.000004
Distribute row-wise data (scalable)                                         |     1  0.000019  0.000019
GPS: create_level_structure                                                 |     4  0.000006  0.000023
Gibbs-Poole-Stockmeyer ordering                                             |     1  0.000058  0.000058
Init dofmap from element dofmap                                             |     4  0.000014  0.000057
SparsityPattern::finalize                                                   |    14  0.000286  0.004011
Topology: create                                                            |     1  0.000258  0.000258
Topology: determine shared index ownership                                  |     1  0.000071  0.000071
Topology: determine vertex ownership groups (owned, undetermined, unowned)  |     1  0.000013  0.000013
~First Order: AltMin-Damage solver                                          |     4  0.000101  0.000405
~First Order: AltMin-Elastic solver                                         |     4  0.002403  0.009611
~Second Order: Bifurcation                                                  |     4  0.260236  1.040944
  0 SNES Function norm 3.909186116588e-01
  1 SNES Function norm 3.271666527291e-01
  2 SNES Function norm 1.996627348698e-01
  3 SNES Function norm 6.294411955572e-16
  0 SNES Function norm 0.000000000000e+00
ALTMIN - Iterations:   1,            Error: 0.0000e+00, alpha_H1,            alpha_max: 0.0000e+00
Saving the eigenvectors for the following eigenvalues
[]
State is elastic: True
State's inertia: (0, 0, 720)
Path is unique: True

[MPI_MAX] Summary of timings                                                |  reps  wall avg  wall tot
-------------------------------------------------------------------------------------------------------
Build dofmap data                                                           |     4  0.000158  0.000633
Build sparsity                                                              |    62  0.000036  0.002217
Compute connectivity 1-2                                                    |     1  0.000004  0.000004
Compute dof reordering map                                                  |     4  0.000004  0.000018
Compute entities of dim = 1                                                 |     1  0.000249  0.000249
Compute local part of mesh dual graph                                       |     1  0.000173  0.000173
Compute local-to-local map                                                  |     1  0.000018  0.000018
Compute-local-to-global links for global/local adjacency list               |     1  0.000004  0.000004
Distribute row-wise data (scalable)                                         |     1  0.000019  0.000019
GPS: create_level_structure                                                 |     4  0.000006  0.000023
Gibbs-Poole-Stockmeyer ordering                                             |     1  0.000058  0.000058
Init dofmap from element dofmap                                             |     4  0.000014  0.000057
SparsityPattern::finalize                                                   |    17  0.000292  0.004967
Topology: create                                                            |     1  0.000258  0.000258
Topology: determine shared index ownership                                  |     1  0.000071  0.000071
Topology: determine vertex ownership groups (owned, undetermined, unowned)  |     1  0.000013  0.000013
~First Order: AltMin-Damage solver                                          |     5  0.000099  0.000497
~First Order: AltMin-Elastic solver                                         |     5  0.002545  0.012723
~Second Order: Bifurcation                                                  |     5  0.214719  1.073597
  0 SNES Function norm 3.909186116588e-01
  1 SNES Function norm 3.271666527291e-01
  2 SNES Function norm 1.996627348698e-01
  3 SNES Function norm 8.560821404371e-16
  0 SNES Function norm 0.000000000000e+00
ALTMIN - Iterations:   1,            Error: 0.0000e+00, alpha_H1,            alpha_max: 0.0000e+00
Saving the eigenvectors for the following eigenvalues
[]
State is elastic: True
State's inertia: (0, 0, 720)
Path is unique: True

[MPI_MAX] Summary of timings                                                |  reps  wall avg  wall tot
-------------------------------------------------------------------------------------------------------
Build dofmap data                                                           |     4  0.000158  0.000633
Build sparsity                                                              |    74  0.000036  0.002657
Compute connectivity 1-2                                                    |     1  0.000004  0.000004
Compute dof reordering map                                                  |     4  0.000004  0.000018
Compute entities of dim = 1                                                 |     1  0.000249  0.000249
Compute local part of mesh dual graph                                       |     1  0.000173  0.000173
Compute local-to-local map                                                  |     1  0.000018  0.000018
Compute-local-to-global links for global/local adjacency list               |     1  0.000004  0.000004
Distribute row-wise data (scalable)                                         |     1  0.000019  0.000019
GPS: create_level_structure                                                 |     4  0.000006  0.000023
Gibbs-Poole-Stockmeyer ordering                                             |     1  0.000058  0.000058
Init dofmap from element dofmap                                             |     4  0.000014  0.000057
SparsityPattern::finalize                                                   |    20  0.000295  0.005894
Topology: create                                                            |     1  0.000258  0.000258
Topology: determine shared index ownership                                  |     1  0.000071  0.000071
Topology: determine vertex ownership groups (owned, undetermined, unowned)  |     1  0.000013  0.000013
~First Order: AltMin-Damage solver                                          |     6  0.000097  0.000579
~First Order: AltMin-Elastic solver                                         |     6  0.002596  0.015577
~Second Order: Bifurcation                                                  |     6  0.184332  1.105992
  0 SNES Function norm 3.909186116588e-01
  1 SNES Function norm 3.271666527291e-01
  2 SNES Function norm 1.996627348698e-01
  3 SNES Function norm 9.779611323703e-16
  0 SNES Function norm 0.000000000000e+00
ALTMIN - Iterations:   1,            Error: 0.0000e+00, alpha_H1,            alpha_max: 0.0000e+00
Saving the eigenvectors for the following eigenvalues
[]
State is elastic: True
State's inertia: (0, 0, 720)
Path is unique: True

[MPI_MAX] Summary of timings                                                |  reps  wall avg  wall tot
-------------------------------------------------------------------------------------------------------
Build dofmap data                                                           |     4  0.000158  0.000633
Build sparsity                                                              |    86  0.000036  0.003094
Compute connectivity 1-2                                                    |     1  0.000004  0.000004
Compute dof reordering map                                                  |     4  0.000004  0.000018
Compute entities of dim = 1                                                 |     1  0.000249  0.000249
Compute local part of mesh dual graph                                       |     1  0.000173  0.000173
Compute local-to-local map                                                  |     1  0.000018  0.000018
Compute-local-to-global links for global/local adjacency list               |     1  0.000004  0.000004
Distribute row-wise data (scalable)                                         |     1  0.000019  0.000019
GPS: create_level_structure                                                 |     4  0.000006  0.000023
Gibbs-Poole-Stockmeyer ordering                                             |     1  0.000058  0.000058
Init dofmap from element dofmap                                             |     4  0.000014  0.000057
SparsityPattern::finalize                                                   |    23  0.000297  0.006824
Topology: create                                                            |     1  0.000258  0.000258
Topology: determine shared index ownership                                  |     1  0.000071  0.000071
Topology: determine vertex ownership groups (owned, undetermined, unowned)  |     1  0.000013  0.000013
~First Order: AltMin-Damage solver                                          |     7  0.000095  0.000668
~First Order: AltMin-Elastic solver                                         |     7  0.002638  0.018465
~Second Order: Bifurcation                                                  |     7  0.162586  1.138104
  0 SNES Function norm 3.909186116588e-01
  1 SNES Function norm 3.271666527291e-01
  2 SNES Function norm 1.996627348698e-01
  3 SNES Function norm 1.115587072823e-15
  0 SNES Function norm 0.000000000000e+00
ALTMIN - Iterations:   1,            Error: 0.0000e+00, alpha_H1,            alpha_max: 0.0000e+00
Saving the eigenvectors for the following eigenvalues
[]
State is elastic: True
State's inertia: (0, 0, 720)
Path is unique: True

[MPI_MAX] Summary of timings                                                |  reps  wall avg  wall tot
[…]
