### 2024-07-10 kumiori/irrevolutions#56 @jorgensd
I also see no good motivation to put all the tests inside `src/irrevolutions`, as I would rather have them at the root of the repository.
