### 2024-07-10 kumiori/irrevolutions#55 @jorgensd
Additionally, what specifically makes it hard to use the `ghcr.io/fenics/dolfinx/dolfinx:v0.7.2` images (out of curiosity as a DOLFINx developer).
