### 2024-07-10 kumiori/irrevolutions#56 @jorgensd
Additionally, I would advice isolating tests inside functions with the prefix `test_*` within each test file as several test does not adhere to this principle
