### 2024-07-28 7f63859
fix and simplify tests (fixtures and output)
