+++ src/irrevolutions/utils/__init__.py
@@ -12,6 +12,7 @@ import yaml
 from dolfinx.fem import assemble_scalar, form
 from mpi4py import MPI
 from petsc4py import PETSc
+import pickle
 
 comm = MPI.COMM_WORLD
 
@@ -380,3 +381,125 @@ def indicator_function(v):
         mode=PETSc.ScatterMode.FORWARD)
     
     return w
+
+
+def save_binary_data(filename, data):
+    viewer = PETSc.Viewer().createBinary(filename, "w")
+
+    if isinstance(data, list):
+        for item in data:
+            item.view(viewer)
+    elif isinstance(data, PETSc.Mat):
+        data.view(viewer)
+    elif isinstance(data, PEtest_binarydataioTSc.Vec):
+        data.view(viewer)
+    else:
+        raise ValueError("Unsupported data type for saving")
+
+
+def load_binary_data(filename):
+    viewer = PETSc.Viewer().createBinary(filename, "r")
+    data = []
+    vectors = []
+
+    i = 0
+    while True:
+        try:
+            vec = PETSc.Vec().load(viewer)
+
+            vectors.append(vec)
+            i += 1
+        except PETSc.Error as e:
+            # __import__('pdb').set_trace()
+            # if e.ierr == PETSc.Error.S_ARG_WRONG:
+            print(f"Error {e.ierr}: {translatePETScERROR.get(e.ierr, 'Unknown error')}")
+            break
+            # else:
+            # raise
+
+    return data
+
+
+def load_binary_vector(filename):
+    """
+    Load a binary file containing a PETSc vector.
+
+    Args:
+        filename (str): Path to the binary file.
+
+    Returns:
+        PETSc.Vec: Loaded PETSc vector.
+    """
+    try:
+        # Create a PETSc viewer for reading
+        viewer = PETSc.Viewer().createBinary(filename, "r")
+
+        # Load the vector from the viewer
+        vector = PETSc.Vec().load(viewer)
+
+        # Close the viewer
+        viewer.destroy()
+
+        return vector
+
+    except PETSc.Error as e:
+        print(f"Error: {e}")
+        return None
+
+
+def load_binary_matrix(filename):
+    """
+    Load a binary file containing a PETSc Matrix.
+
+    Args:
+        filename (str): Path to the binary file.
+
+    Returns:
+        PETSc.Mat: Loaded PETSc Matrix.
+    """
+    try:
+        # Create a PETSc viewer for reading
+        viewer = PETSc.Viewer().createBinary(filename, "r")
+
+        # Load the vector from the viewer
+        vector = PETSc.Mat().load(viewer)
+
+        # Close the viewer
+        viewer.destroy()
+
+        return vector
+
+    except PETSc.Error as e:
+        print(f"Error: {e}")
+        return None
+
+
+def save_minimal_constraints(obj, filename):
+    minimal_constraints = {
+        "bglobal_dofs_mat": obj.bglobal_dofs_mat,
+        "bglobal_dofs_mat_stacked": obj.bglobal_dofs_mat_stacked,
+        "bglobal_dofs_vec": obj.bglobal_dofs_vec,
+        "bglobal_dofs_vec_stacked": obj.bglobal_dofs_vec_stacked,
+        "blocal_dofs": obj.blocal_dofs,
+        "boffsets_mat": obj.boffsets_mat,
+        "boffsets_vec": obj.boffsets_vec,
+    }
+
+    with open(filename, "wb") as file:
+        pickle.dump(minimal_constraints, file)
+
+
+def load_minimal_constraints(filename):
+    import irrevolutions.solvers.restriction as restriction
+    
+    with open(filename, "rb") as file:
+        minimal_constraints = pickle.load(file)
+
[…]
