+++ src/irrevolutions/test/test_1d.py
@@ -1,747 +1,0 @@
-#!/usr/bin/env python3
-import json
-import logging
-import os
-import sys
-from pathlib import Path
-
-import dolfinx
-import dolfinx.mesh
-import dolfinx.plot
-import numpy as np
-import pandas as pd
-import petsc4py
-import pyvista
-import ufl
-import yaml
-from dolfinx.fem import (
-    Constant,
-    Function,
-    assemble_scalar,
-    dirichletbc,
-    form,
-    locate_dofs_geometrical,
-    set_bc,
-)
-from dolfinx.common import list_timings
-from dolfinx.fem.petsc import assemble_vector, set_bc
-from dolfinx.io import XDMFFile
-from mpi4py import MPI
-from petsc4py import PETSc
-
-from irrevolutions.algorithms.am import HybridSolver
-from irrevolutions.algorithms.so import BifurcationSolver, StabilitySolver
-from irrevolutions.solvers import SNESSolver
-from irrevolutions.solvers.function import vec_to_functions
-from irrevolutions.utils import (
-    ColorPrint,
-    _logger,
-    _write_history_data,
-    history_data,
-    norm_H1,
-    norm_L2,
-)
-from irrevolutions.utils.plots import (
-    plot_AMit_load,
-    plot_energies,
-    plot_force_displacement,
-)
-
-#
-from irrevolutions.utils.viz import plot_profile
-
-"""The fundamental problem of a 1d bar in traction.
-0|(WWWWWWWWWWWWWWWWWWWWWW)|========> t
-
-(WWW): endommageable bar, y=(u, alpha)
-load: displacement hard-t
-"""
-
-
-# logging.getLogger().setLevel(logging.INFO)
-
-
-class _AlternateMinimisation1D:
-
-    def __init__(
-        self,
-        total_energy,
-        state,
-        bcs,
-        solver_parameters={},
-        bounds=(dolfinx.fem.function.Function, dolfinx.fem.function.Function),
-    ):
-        self.state = state
-        self.alpha = state["alpha"]
-        self.alpha_old = dolfinx.fem.function.Function(self.alpha.function_space)
-        self.u = state["u"]
-        self.alpha_lb = bounds[0]
-        self.alpha_ub = bounds[1]
-        self.total_energy = total_energy
-        self.solver_parameters = solver_parameters
-
-        V_u = state["u"].function_space
-        V_alpha = state["alpha"].function_space
-
-        energy_u = ufl.derivative(self.total_energy, self.u, ufl.TestFunction(V_u))
-        energy_alpha = ufl.derivative(
-            self.total_energy, self.alpha, ufl.TestFunction(V_alpha)
-        )
-
-        self.F = [energy_u, energy_alpha]
-
-        self.elasticity = SNESSolver(
-            energy_u,
-            self.u,
-            bcs.get("bcs_u"),
-            bounds=None,
-            petsc_options=self.solver_parameters.get("elasticity").get("snes"),
-            prefix=self.solver_parameters.get("elasticity").get("prefix"),
-        )
-
-        self.damage = SNESSolver(
-            energy_alpha,
-            self.alpha,
-            bcs.get("bcs_alpha"),
-            bounds=(self.alpha_lb, self.alpha_ub),
-            petsc_options=self.solver_parameters.get("damage").get("snes"),
-            prefix=self.solver_parameters.get("damage").get("prefix"),
-        )
-
-    def solve(self, outdir=None):
-
-        alpha_diff = dolfinx.fem.Function(self.alpha.function_space)
-
-        self.data = {
-            "iteration": [],
-            "error_alpha_L2": [],
-            "error_alpha_H1": [],
-            "F_norm": [],
[…]
