+++ src/irrevolutions/test/test_bifurcation.py
@@ -1,239 +1,0 @@
-#!/usr/bin/env python3
-import pandas as pd
-from irrevolutions.utils import ColorPrint
-from irrevolutions.utils.plots import plot_energies
-from irrevolutions.meshes.primitives import mesh_bar_gmshapi
-from irrevolutions.algorithms.so import BifurcationSolver
-from irrevolutions.algorithms.am import AlternateMinimisation
-from irrevolutions.models import DamageElasticityModel as Brittle
-import dolfinx.mesh
-from dolfinx.fem import (
-    Constant,
-    Function,
-    FunctionSpace,
-    assemble_scalar,
-    dirichletbc,
-    form,
-    locate_dofs_geometrical,
-    set_bc,
-)
-from dolfinx.io import XDMFFile, gmshio
-import numpy as np
-import yaml
-import json
-from pathlib import Path
-import sys
-import os
-from mpi4py import MPI
-import petsc4py
-from petsc4py import PETSc
-import dolfinx
-import dolfinx.plot
-import ufl
-import numpy as np
-from dolfinx.common import list_timings
-
-import logging
-
-logging.basicConfig(level=logging.INFO)
-
-
-# ///////////
-
-
-petsc4py.init(sys.argv)
-comm = MPI.COMM_WORLD
-
-# Mesh on node model_rank and then distribute
-model_rank = 0
-
-
-with open(os.path.join(os.path.dirname(__file__), "parameters.yml")) as f:
-    parameters = yaml.load(f, Loader=yaml.FullLoader)
-
-# Get mesh parameters
-Lx = parameters["geometry"]["Lx"]
-Ly = parameters["geometry"]["Ly"]
-tdim = parameters["geometry"]["geometric_dimension"]
-_nameExp = parameters["geometry"]["geom_type"]
-ell_ = parameters["model"]["ell"]
-lc = ell_ / 5.0
-
-
-# Get geometry model
-geom_type = parameters["geometry"]["geom_type"]
-
-# Create the mesh of the specimen with given dimensions
-gmsh_model, tdim = mesh_bar_gmshapi(geom_type, Lx, Ly, lc, tdim)
-
-# Get mesh and meshtags
-mesh, mts, fts = gmshio.model_to_mesh(gmsh_model, comm, model_rank, tdim)
-
-
-outdir = os.path.join(os.path.dirname(__file__), "output")
-prefix = os.path.join(outdir, "traction")
-if comm.rank == 0:
-    Path(prefix).mkdir(parents=True, exist_ok=True)
-
-
-with XDMFFile(
-    comm, f"{prefix}/{_nameExp}.xdmf", "w", encoding=XDMFFile.Encoding.HDF5
-) as file:
-    file.write_mesh(mesh)
-
-# Function spaces
-element_u = ufl.VectorElement("Lagrange", mesh.ufl_cell(), degree=1, dim=tdim)
-V_u = FunctionSpace(mesh, element_u)
-
-element_alpha = ufl.FiniteElement("Lagrange", mesh.ufl_cell(), degree=1)
-V_alpha = FunctionSpace(mesh, element_alpha)
-
-# Define the state
-u = Function(V_u, name="Displacement")
-u_ = Function(V_u, name="Boundary Displacement")
-zero_u = Function(V_u, name="   Boundary Displacement")
-alpha = Function(V_alpha, name="Damage")
-zero_alpha = Function(V_alpha, name="Damage Boundary Field")
-
-state = {"u": u, "alpha": alpha}
-
-# need upper/lower bound for the damage field
[…]
