### 2024-07-29 abdd80c
moving up tests in dirtree
