+++ pytest.ini
@@ -1,5 +1,5 @@
 [pytest]
-addopts = -v --ignore=test_*
+addopts = -v --cov-report=term
 
 [pytest-cov]
 omit = src/test/*
