+++ README.md
@@ -45,6 +45,13 @@ On a windows box:
 docker run --rm -ti -v "C:/...":/home/numerix" -w /home/numerix kumiori3\numerix:latest
 ```
 
+To install the software, run DOLFINx through the aforementioned docker container and install the irrevolutions-package with 
+
+```python3 -m pip install .```
+
+from the root of this repository
+
+
 This code was initially conceived as a support for the teaching course MEC647, 
 (Complex) Crack Propagation in Brittle Materials, delivered to the students of the international
 master programme, joint between École Polytechnique and ENSTA-Paristech throughout 2020-2022. 
