+++ .github/workflows/pull-image.yml
@@ -24,19 +24,19 @@ jobs:
         run: |
           docker run --rm \
             -v ${{ github.workspace }}:/workspace kumiori3/numerix:stable-amd64 \
-            # sh -c "cd /workspace && ls -la && pip install . && python3 -c 'import irrevolutions' && cd src/irrevolutions/test/ && pytest -v ."
+            # sh -c "cd /workspace && ls -la && pip install . && python3 -c 'import irrevolutions' && pytest -v ."
 
       - name: Run Docker Container & Pytest Simple Script
         run: |
             docker run --rm \
             -v ${{ github.workspace }}:/workspace kumiori3/numerix:stable-amd64 \
-            # sh -c "cd /workspace && ls -la && pip install . && python3 -c 'import irrevolutions' && cd src/irrevolutions/test/ && python3 -m pytest test_rayleigh.py"
+            # sh -c "cd /workspace && ls -la && pip install . && python3 -c 'import irrevolutions' && cd test/ && python3 -m pytest test_rayleigh.py"
 
       - name: Run Docker Container & Run Simple Script
         run: |
             docker run --rm \
             -v ${{ github.workspace }}:/workspace kumiori3/numerix:stable-amd64 \
-            sh -c "cd /workspace && ls -la && pip install . && python3 -c 'import irrevolutions' && cd src/irrevolutions/test/ && python3 test_rayleigh.py"
+            sh -c "cd /workspace && ls -la && pip install . && python3 -c 'import irrevolutions' && cd test/ && python3 test_rayleigh.py"
 
     
     # - name: Install package
