### 2024-07-29 5be0820
simple postprocessing tests
