### 2024-07-29 8035be7
fixed read in parallel
