### 2024-07-30 kumiori/irrevolutions#54 @kumiori
Thank for pointing out.

The commits you refer to (e.g. by SarahTACHET, ayoubrous, but also jaedong2019, mbilinsky, ...) are by students of an academic course taught for Polytechnique/ENSTA for which this code has been initially developed. These commits represent homework assignments. The commits by d2f4d131d9eac6cc27e3d6245ab1476b are from an alternative account of mine.

On the other hand, Pierluigi Cesana, co-author of the JOSS submission, has collaborated on the construction of the algorithms these codes are based upon; has written and tested parts of the pseudo-codes; has contributed in the analysis of the well-posedness of the mathematical problem; has collaborated to the conception of the (analytical and numerical) examples as well as to writing the short accompanying paper, however he was not involved in the actual coding.
