+++ .github/workflows/pull-image.yml
@@ -38,7 +38,13 @@ jobs:
             -v ${{ github.workspace }}:/workspace kumiori3/numerix:stable-amd64 \
             sh -c "cd /workspace && ls -la && pip install . && python3 -c 'import irrevolutions' && cd test/ && python3 test_rayleigh.py"
 
-    
+
+      - name: Run Docker Container & Run pytest
+        run: |
+            docker run --rm \
+            -v ${{ github.workspace }}:/workspace kumiori3/numerix:stable-amd64 \
+            sh -c "cd /workspace && ls -la && pip install . && python3 -c 'import irrevolutions' && pytest -v test/"
+          
     # - name: Install package
     #     run: |
     #         ls -la
