### 2024-07-30 56190a2
simplifying workflow (thanks @jorgensd)
