+++ demo/demo_elasticity.py
@@ -1,0 +1,173 @@
+#!/usr/bin/env python3
+import pyvista
+import dolfinx.mesh
+from dolfinx.io import XDMFFile, gmshio
+import numpy as np
+import yaml
+import json
+from pathlib import Path
+import sys
+import os
+from mpi4py import MPI
+import petsc4py
+from petsc4py import PETSc
+import dolfinx
+import dolfinx.plot
+from dolfinx import log
+import ufl
+
+from pyvista.utilities import xvfb
+
+from irrevolutions.models import ElasticityModel
+from irrevolutions.solvers import SNESSolver as ElasticitySolver
+from irrevolutions.meshes.primitives import mesh_bar_gmshapi
+from irrevolutions.utils.viz import plot_vector
+
+import numpy as np
+import logging
+
+logging.basicConfig(level=logging.INFO)
+
+
+# ///////////
+
+
+petsc4py.init(sys.argv)
+log.set_log_level(log.LogLevel.WARNING)
+
+comm = MPI.COMM_WORLD
+
+# Mesh on node model_rank and then distribute
+model_rank = 0
+
+with open(os.path.join(os.path.dirname(__file__), "parameters.yml")) as f:
+    parameters = yaml.load(f, Loader=yaml.FullLoader)
+
+# Get mesh parameters
+Lx = parameters["geometry"]["Lx"]
+Ly = parameters["geometry"]["Ly"]
+tdim = parameters["geometry"]["geometric_dimension"]
+lc = parameters["geometry"]["lc"]
+
+# Get geometry model
+geom_type = parameters["geometry"]["geom_type"]
+
+# Create the mesh of the specimen with given dimensions
+gmsh_model, tdim = mesh_bar_gmshapi(geom_type, Lx, Ly, lc, tdim)
+
+# Get mesh and meshtags
+mesh, mts, fts = gmshio.model_to_mesh(gmsh_model, comm, model_rank, tdim)
+
+outdir = os.path.join(os.path.dirname(__file__), "output")
+if comm.rank == 0:
+    Path(outdir).mkdir(parents=True, exist_ok=True)
+
+prefix = os.path.join(outdir, "elasticity")
+
+with XDMFFile(comm, f"{prefix}.xdmf", "w", encoding=XDMFFile.Encoding.HDF5) as file:
+    file.write_mesh(mesh)
+
+# Function spaces
+element_u = ufl.VectorElement("Lagrange", mesh.ufl_cell(), degree=1, dim=tdim)
+V_u = dolfinx.fem.FunctionSpace(mesh, element_u)
+V_ux = dolfinx.fem.FunctionSpace(
+    mesh, ufl.FiniteElement("Lagrange", mesh.ufl_cell(), degree=1)
+)
+
+# Define the state
+u = dolfinx.fem.Function(V_u, name="Displacement")
+u_ = dolfinx.fem.Function(V_u, name="Boundary Displacement")
+ux_ = dolfinx.fem.Function(V_ux, name="Boundary Displacement")
+zero_u = dolfinx.fem.Function(V_u, name="   Boundary Displacement")
+
+state = {"u": u}
+
+# Measures
+dx = ufl.Measure("dx", domain=mesh)
+ds = ufl.Measure("ds", domain=mesh)
+
+dofs_u_left = dolfinx.fem.locate_dofs_geometrical(V_u, lambda x: np.isclose(x[0], 0.0))
+dofs_u_right = dolfinx.fem.locate_dofs_geometrical(V_u, lambda x: np.isclose(x[0], Lx))
+dofs_ux_right = dolfinx.fem.locate_dofs_geometrical(
[…]
