### 2024-07-30 e759256
simplyfing CI workflows
