+++ .github/workflows/pull-image.yml
@@ -1,61 +1,0 @@
-name: Run Tests in Docker Container
-run-name: ${{ github.actor }} is pulling containers
-
-on:
-  push:
-    branches:
-      - '*'
-
-jobs:
-  set-up-computing-environment:
-    runs-on: ubuntu-latest
-
-    steps:
-      # Step 0: Checkout
-      - name: Checkout repository
-        uses: actions/checkout@v4
-      
-      # Step 1: Pull the Docker container from Docker Hub
-      - name: Pull Docker Container
-        run: docker pull kumiori3/numerix:stable-amd64
-
-      # Step 2: Check out the repository code within the container
-      - name: Run Docker Container & Test
-        run: |
-          docker run --rm \
-            -v ${{ github.workspace }}:/workspace kumiori3/numerix:stable-amd64 \
-            # sh -c "cd /workspace && ls -la && pip install . && python3 -c 'import irrevolutions' && pytest -v ."
-
-      - name: Run Docker Container & Pytest Simple Script
-        run: |
-            docker run --rm \
-            -v ${{ github.workspace }}:/workspace kumiori3/numerix:stable-amd64 \
-            # sh -c "cd /workspace && ls -la && pip install . && python3 -c 'import irrevolutions' && cd test/ && python3 -m pytest test_rayleigh.py"
-
-      - name: Run Docker Container & Run Simple Script
-        run: |
-            docker run --rm \
-            -v ${{ github.workspace }}:/workspace kumiori3/numerix:stable-amd64 \
-            sh -c "cd /workspace && ls -la && pip install . && python3 -c 'import irrevolutions' && cd test/ && python3 test_rayleigh.py"
-
-
-      - name: Run Docker Container & Run pytest
-        run: |
-            docker run --rm \
-            -v ${{ github.workspace }}:/workspace kumiori3/numerix:stable-amd64 \
-            sh -c "cd /workspace && ls -la && pip install . && python3 -c 'import irrevolutions' && pytest -v test/"
-          
-    # - name: Install package
-    #     run: |
-    #         ls -la
-    #         pip install .
-
-    #   - name: Test import package
-    #     run: |
-    #       python3 -c "import irrevolutions" && echo "Imported successfully" || echo "Failed to import"
-
-    #   - name: Test scripts
-    #     run: |
-    #       pytest -v tests/
-  
-    
