### 2024-07-30 b745167
moving utility samples
