### 2024-07-30 kumiori/irrevolutions#56 @kumiori
Thank you very much for the comments and suggestions.

Tests have been moved to the root, cleaned, isolated within test_* prefixed functions, and the workflow simplified as suggested. All (10) tests pass albeit with several warnings, currently under investigation.
