+++ paper/paper.md
@@ -37,7 +37,7 @@ We study irreversible evolutionary processes with a general energetic notion of 
 
 # Statement of need
 
-Quasi-static evolution problems arising in fracture are strongly nonlinear [@marigo:2023-la-mecanique], [@bourdin:2008-the-variational]. They can admit multiple solutions, or none [@leon-baldelli:2021-numerical]. This demands both a functional theoretical framework and practical computational tools for real case scenarios. Due to the lack of uniqueness of solutions, it is fundamental to leverage the full variational structure of the problem and investigate up to second order, to detect nucleation of stable modes and transitions of unstable states. The stability of a multiscale system along its nontrivial evolutionary paths in phase space is a key property that is difficult to check: numerically, for real case scenarios with several length scales involved, and analytically, in the infinite-dimensional setting. ~~The current literature in computational fracture mechanics predominantly focuses on unilateral first-order criteria, systematically neglecting the exploration of higher-order information for critical points.~~ **Despite the concept of unilateral stability is classical in the variational theory of irreversible systems \cite{mielke} and the mechanics of fracture \cite{FRANCFORT} (see also \cite{bazant, petryk, nguyen}), few studies have explored second-order criteria for crack nucleation and evolution. Although sporadic, these studies are significant, including \cite{pham:2011-the-issues}, \cite{Pham2013aa}, \cite{SICSIC}, \cite{leon-baldelli:2021-numerical}, and \cite{camilla}. The current literature in computational fracture mechanics predominantly focuses on unilateral first-order criteria, systematically neglecting the exploration of higher-order information for critical points. To the best of our knowledge, no general numerical tools are available to address second-order criteria in evolutionary nonlinear irreversible systems and fracture mechanics.**
+Quasi-static evolution problems arising in fracture are strongly nonlinear [@marigo:2023-la-mecanique], [@bourdin:2008-the-variational]. They can admit multiple solutions, or none [@leon-baldelli:2021-numerical]. This demands both a functional theoretical framework and practical computational tools for real case scenarios. Due to the lack of uniqueness of solutions, it is fundamental to leverage the full variational structure of the problem and investigate up to second order, to detect nucleation of stable modes and transitions of unstable states. The stability of a multiscale system along its nontrivial evolutionary paths in phase space is a key property that is difficult to check: numerically, for real case scenarios with several length scales involved, and analytically, in the infinite-dimensional setting. ~~The current literature in computational fracture mechanics predominantly focuses on unilateral first-order criteria, systematically neglecting the exploration of higher-order information for critical points.~~ **Despite the concept of unilateral stability is classical in the variational theory of irreversible systems \cite{mielke} and the mechanics of fracture \cite{FRANCFORT} (see also \cite{bazant, petryk, nguyen}), few studies have explored second-order criteria for crack nucleation and evolution. Although sporadic, these studies are significant, including \cite{pham:2011-the-issues}, \cite{Pham2013aa}, \cite{SICSIC}, \cite{leon-baldelli:2021-numerical}, and \cite{camilla}.** The current literature in computational fracture mechanics predominantly focuses on unilateral first-order criteria, systematically neglecting the exploration of higher-order information for critical points. **To the best of our knowledge, no general numerical tools are available to address second-order criteria in evolutionary nonlinear irreversible systems and fracture mechanics.**
 
 To fill this gap, our nonlinear solvers offer a flexible toolkit for advanced stability analysis of systems which evolve with constraints.
 
@@ -70,16 +70,16 @@ Above, $T$ defines a horizon of events. The system is represented by its total e
 
[…]
