### 2024-08-02 kumiori/irrevolutions#53 @kumiori
Thank you for your comments, all addressed below.

1. we removed the last sentence as recommended

2. In so-called "phase field" fractures, questions of stability are not common practice. We have expanded our consideration in the text. As a reference, you can take the recent book "MEALOR II Damage Mechanics and Local Approach to Fracture" (doi: 10.5281/zenodo.10125170)

> "the easiest to implement is the phase-field approach due to its variational nature. This explains its current success. A time-discretization is first performed which gives an incremental objective function. This function is then minimized using an alternate minimization. For instance, for elasto-damage models, a basic finite element solver is used to update the displacement field while another basic finite element solver updates the damage. These two solvers alternate until convergence. Then the next time-step is considered."

For other references of peer-reviewed papers, see also:
https://doi.org/10.1007/s10704-024-00763-w
https://doi.org/10.1016/j.cma.2010.04.011

3. the problem specification has been moved prior to introducing the solvers, as suggested
2. the section has been renamed as suggested
3. thank you for pointing out the missing definition of $V_0$, it has been specified in the text
4. we have expanded our references to stability problems driven by unilateral principles, as requested, to elucidate the context for non-specialists
4. we have clarified the use of the bounds which apply to the internal (irreversible) order parameter. Currently, the ordering is implicit, namely, the first element of the list ```bounds``` is the lower bound, the second is the upper bound.

4. we have added, in the appendix, an example of numerical parameters referring to the computation shown in the manuscript
4. indeed, thank you for pointing out, we have specified the argument of the operator.
4. we have added the explicit solution of the Rayleigh ratio minimisation problems in the appendix, for reference
4. we have expanded our comment in relation to other existing numerical software in the Statement of Need. To the best of our knowledge, there are no commonly-used packages for the solution of the general problem we address.
