### 2024-08-02 kumiori/irrevolutions#52 @pcesana81
We have updated the Bibliography file with all missing DOIs. Please notice that some references come without a DOI.
