+++ README.md
@@ -35,14 +35,18 @@ Asymmetrically, feature-work is `rebased`.
 First, run the container, attaching an interactive session and sharing data space 
 (the current dir) between the host and the container (the syntax is origin:target).
 
-On a Mac:
+On an ARM-based machine:
 ```
-docker run --rm -ti -v "$(pwd)":/home/numerix -w /home/numerix kumiori3/numerix:latest
+docker run --rm -ti -v "$(pwd)":/home/numerix -w /home/numerix kumiori3/numerix:stable
+```
+On an AMD64 machine:
+```
+docker run --rm -ti -v "$(pwd)":/home/numerix -w /home/numerix kumiori3/numerix:stable-amd64
 ```
 
 On a windows box:
 ```
-docker run --rm -ti -v "C:/...":/home/numerix" -w /home/numerix kumiori3\numerix:latest
+docker run --rm -ti -v "C:/...":/home/numerix" -w /home/numerix kumiori3\numerix:stable-amd64
 ```
 
 To install the software, run DOLFINx through the aforementioned docker container and install the irrevolutions-package with 
