### 2024-08-02 96f13a8
readme and project updates
