### 2024-08-02 kumiori/irrevolutions#55 @kumiori
Thank you for the suggestions.

The README.md has been updated with installation instructions and the correct container images,
the toml file is now pinned to the fenics-dolfinx version as suggested.

We target a custom docker container built upon the image you refer to because it contains the necessary software for postprocessing with pyvista, which is not strictly required.

We may suggest a lighter software installation setup by targeting the base DOLFINx images. Do you think this would be appropriate?
