### 2024-08-02 #6897 @kumiori
Thank you for the reaching out. All issues have been addressed and commented. Best

> On 26 Jul 2024, at 01:25, Mark Jensen ***@***.***> wrote:
>
>
> Thanks @jorgensd <https://github.com/jorgensd> - @kumiori <https://github.com/kumiori>, can you please check your repo for the issues raised? 5 open issues here <https://github.com/kumiori/irrevolutions/issues>. Thank you
>
