### 2024-08-04 kumiori/irrevolutions#55 @jorgensd
> We target a custom docker container built upon the image you refer to because it contains the necessary software for postprocessing with pyvista, which is not strictly required.

As far as your tests go, it seems like pyvista is a strong requirement, ref the failing CI:
https://github.com/kumiori/irrevolutions/actions/runs/10214169939/job/28261025672
```bash
test/test_1d.py:33: in <module>
    from irrevolutions.algorithms.so import BifurcationSolver, StabilitySolver
/usr/local/lib/python3.10/dist-packages/irrevolutions/algorithms/so.py:16: in <module>
    import irrevolutions.solvers.slepcblockproblem as eigenblockproblem
/usr/local/lib/python3.10/dist-packages/irrevolutions/solvers/slepcblockproblem.py:9: in <module>
    from irrevolutions.utils.viz import plot_matrix
/usr/local/lib/python3.10/dist-packages/irrevolutions/utils/viz.py:23: in <module>
    xvfb.start_xvfb(wait=0.05)
/usr/local/lib/python3.10/dist-packages/pyvista/plotting/utilities/xvfb.py:50: in start_xvfb
    raise OSError(XVFB_INSTALL_NOTES)
E   OSError: Please install Xvfb with:
E
E   Debian
E   $ sudo apt install libgl1-mesa-glx xvfb
```

Secondly, the aforementioned image: `ghcr.io/fenics/dolfinx/dolfinx:v0.7.2` does not contain pyvista.
You could use `ghcr.io/fenics/dolfinx/lab:v0.7.2` if you require pyvista as part of your software.
That image can be executed as any normal image with the following command:
```bash
docker run -ti --rm --entrypoint=/bin/bash ghcr.io/fenics/dolfinx/lab:v0.7.3
```
i.e. changing the entrypoint of the container to bash rather than `jupyter lab ....`
