### 2024-08-04 kumiori/irrevolutions#53 @jorgensd
@kumiori The changes look really good.
Some minor typos:
```
of the system as a dictionary (u, alpha), whre the first elemnt is associated to the reversible
field and the second to the irreversible component, its the associated constraints on the latter,
and the solver’s parameters (see an example in the Appendix).
```
Should be
```
of the system as a dictionary (u, alpha), **where** the first **element** is associated to the reversible
field and the second to the irreversible component, its the associated constraints on the latter,
and the solver’s parameters (see an example in the Appendix).
