### 2024-08-04 kumiori/irrevolutions#53 @jorgensd
In the appendix, could one format the three different conditions (1., 2., 3.) in a slightly more readable way?
I get that there are a lot of conditionals for each $\beta$ solution, but maybe using sub-equations (4.1, 4.2 and 4.3 maybe)?

Additionally, eq [2.1] and [2.2] is not explicitly defined, only through the [2] with an implicit assumption that the "and-statement" gives the sub-numbering. I am not super fond of this, but it shouldn't be a blocker to proceed further with the paper..
