### 2024-08-04 kumiori/irrevolutions#52 @jorgensd
Capitalization of following papers should be adjusted according to what the journal has published:
- Stable States and Paths of Stmuctures with Plasticity or Damage
- Stability of Homogeneous States with Gradient Damage Models: Size Effects and Shape Effects in the Three-Dimensional Setting
- Cone-constrained eigenvalue problems: theory and algorithms
- Parallel distributed computing using Python
