### 2024-08-04 kumiori/irrevolutions#64 @jorgensd
Document issue submission and support
Minor issue, to comply with:
**Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support**
in https://github.com/openjournals/joss-reviews/issues/6897
Add a sentence referencing the issue tracker and any other means of communication in the README.md
