### 2024-08-04 kumiori/irrevolutions#65 @jorgensd
Reproduce figures from paper
What scripts would I have to run to reproduce figures 1-5 in the paper?

This is to comply with

> Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.

from https://github.com/openjournals/joss-reviews/issues/6897
Would it be sufficient to run `playground/nb/rayleigh_benchmark.ipynb`?

Seems like the notebook is a bit outdated, as it uses `import eigenspace as eig` which should probably be `from irrevolutions.utils import eigenspace as eig`.
