### 2024-08-07 kumiori/irrevolutions#66 @jhale
Paper issues
- Petryk and Nguyen references missing (cannot see references at all on latest generated version, infact).
- Bold is overused in my view
- Typo line 83
- In the example of use it says that E takes a `dolfinx.fem.form` but actually the classes take `ufl.form.Form`.
- The paper is quite jargon heavy throughout, may be difficult to understand for a more general audience (even mechanics and applied maths!).

https://github.com/openjournals/joss-reviews/issues/6897
