### 2024-08-07 kumiori/irrevolutions#67 @jhale
DOLFINy
The classes SLEPcBlockProblem are very close to those in DOLFINy to the extent that the Copyright of those files should remain with the original authors and the license should be included (LGPLv3). The README should note that it contains content from DOLFINy.

https://github.com/openjournals/joss-reviews/issues/6897
