### 2024-08-07 kumiori/irrevolutions#68 @jhale
Contributing file
To tick the box please add something like a CONTRIBUTING.md file, e.g.:

https://github.com/FEniCS/dolfinx/blob/main/CONTRIBUTING.md

https://github.com/openjournals/joss-reviews/issues/6897
