### 2024-08-07 kumiori/irrevolutions#69 @jhale
README
The first sentence of the README is unclear. Please just provide a clear and concise explanation of what the package does.

https://github.com/openjournals/joss-reviews/issues/6897
