### 2024-08-13 6726e13
fixed issue #62; thanks Jorgen
