+++ src/irrevolutions/practice/discrete_atk.py
@@ -30,10 +30,7 @@ import dolfinx.plot
 from dolfinx import log
 import ufl
 
-from dolfinx.fem.petsc import (
-    set_bc,
-    assemble_vector
-    )
+from dolfinx.fem.petsc import set_bc, assemble_vector
 from dolfinx.io import XDMFFile, gmshio
 import logging
 from dolfinx.common import Timer, list_timings, TimingType
@@ -45,8 +42,6 @@ from meshes.primitives import mesh_bar_gmshapi
 from irrevolutions.utils import ColorPrint
 from utils.plots import plot_energies
 from irrevolutions.utils import norm_H1, norm_L2
-
-
 
 
 sys.path.append("../")
@@ -65,19 +60,21 @@ load: displacement hard-t
 
 
 from solvers.function import functions_to_vec
+
 logging.getLogger().setLevel(logging.CRITICAL)
 
 comm = MPI.COMM_WORLD
 
+
 class _AlternateMinimisation:
-    def __init__(self,
-                total_energy,
-                state,
-                bcs,
-                solver_parameters={},
-                bounds=(dolfinx.fem.function.Function,
-                        dolfinx.fem.function.Function)
-                ):
+    def __init__(
+        self,
+        total_energy,
+        state,
+        bcs,
+        solver_parameters={},
+        bounds=(dolfinx.fem.function.Function, dolfinx.fem.function.Function),
+    ):
         self.state = state
         self.alpha = state["alpha"]
         self.alpha_old = dolfinx.fem.function.Function(self.alpha.function_space)
@@ -90,8 +87,7 @@ class _AlternateMinimisation:
         V_u = state["u"].function_space
         V_alpha = state["alpha"].function_space
 
-        energy_u = ufl.derivative(
-            self.total_energy, self.u, ufl.TestFunction(V_u))
+        energy_u = ufl.derivative(self.total_energy, self.u, ufl.TestFunction(V_u))
         energy_alpha = ufl.derivative(
             self.total_energy, self.alpha, ufl.TestFunction(V_alpha)
         )
@@ -117,7 +113,6 @@ class _AlternateMinimisation:
         )
 
     def solve(self, outdir=None):
-
         alpha_diff = dolfinx.fem.Function(self.alpha.function_space)
 
         self.data = {
@@ -154,10 +149,7 @@ class _AlternateMinimisation:
             Fv = [assemble_vector(form(F)) for F in self.F]
 
             Fnorm = np.sqrt(
-                np.array(
-                    [comm.allreduce(Fvi.norm(), op=MPI.SUM)
-                        for Fvi in Fv]
-                ).sum()
+                np.array([comm.allreduce(Fvi.norm(), op=MPI.SUM) for Fvi in Fv]).sum()
             )
 
             error_alpha_max = alpha_diff.vector.max()[1]
@@ -204,10 +196,8 @@ class _AlternateMinimisation:
             self.data["solver_u_it"].append(solver_u_it)
             self.data["total_energy"].append(total_energy_int)
 
-
             if (
-                self.solver_parameters.get(
-                    "damage_elasticity").get("criterion")
+                self.solver_parameters.get("damage_elasticity").get("criterion")
                 == "residual_u"
             ):
                 if error_residual_F <= self.solver_parameters.get(
@@ -215,8 +205,7 @@ class _AlternateMinimisation:
                 ).get("alpha_rtol"):
                     break
             if (
[…]
