+++ playground/benchmark-umut-at2/vs_analytics_at2.py
@@ -17,8 +17,15 @@ import pyvista
 import ufl
 import yaml
 from dolfinx.common import list_timings
-from dolfinx.fem import (Constant, Function, assemble_scalar, dirichletbc,
-                         form, locate_dofs_geometrical, set_bc)
+from dolfinx.fem import (
+    Constant,
+    Function,
+    assemble_scalar,
+    dirichletbc,
+    form,
+    locate_dofs_geometrical,
+    set_bc,
+)
 from dolfinx.fem.petsc import assemble_vector, set_bc
 from dolfinx.io import XDMFFile
 from irrevolutions.algorithms.am import HybridSolver
@@ -26,13 +33,22 @@ from irrevolutions.algorithms.so import BifurcationSolver, StabilitySolver
 from irrevolutions.solvers import SNESSolver
 from irrevolutions.solvers.function import vec_to_functions
 from irrevolutions.test.test_1d import _AlternateMinimisation1D as am1d
-from irrevolutions.utils import (ColorPrint, ResultsStorage, Visualization,
-                                 _logger, _write_history_data, history_data,
-                                 norm_H1, norm_L2)
-from irrevolutions.utils.plots import (plot_AMit_load, plot_energies,
-                                       plot_force_displacement)
-from irrevolutions.utils.viz import (plot_mesh, plot_profile, plot_scalar,
-                                     plot_vector)
+from irrevolutions.utils import (
+    ColorPrint,
+    ResultsStorage,
+    Visualization,
+    _logger,
+    _write_history_data,
+    history_data,
+    norm_H1,
+    norm_L2,
+)
+from irrevolutions.utils.plots import (
+    plot_AMit_load,
+    plot_energies,
+    plot_force_displacement,
+)
+from irrevolutions.utils.viz import plot_mesh, plot_profile, plot_scalar, plot_vector
 from mpi4py import MPI
 from petsc4py import PETSc
 from pyvista.utilities import xvfb
@@ -48,6 +64,7 @@ def a(alpha):
     # k_res = parameters["model"]['k_res']
     return (1 - alpha) ** 2
 
+
 def w(alpha):
     """
     Return the homogeneous damage energy term,
@@ -58,8 +75,10 @@ def w(alpha):
     return alpha**2
     # return alpha
 
-def elastic_energy_density(state, 
-                           u_zero: Optional[dolfinx.fem.function.Function] = None):
+
+def elastic_energy_density(
+    state, u_zero: Optional[dolfinx.fem.function.Function] = None
+):
     """
     Returns the elastic energy density of the state.
     """
@@ -70,16 +89,17 @@ def elastic_energy_density(state,
 
     _mu = parameters["model"]["E"]
     _kappa = parameters["model"].get("kappa", 1.0)
-    
+
     # energy_density = _mu / 2.0 * ufl.inner(eps, eps)
     energy_density = _mu / 2.0 * a(alpha) * ufl.inner(eps, eps)
-    
+
     if u_zero is None:
         u_zero = Constant(u.function_space.mesh, 0.0)
 
     substrate_density = _kappa / 2.0 * ufl.inner(u - u_zero, u - u_zero)
 
     return energy_density + substrate_density
+
 
 def damage_energy_density(state):
     """
@@ -93,10 +113,12 @@ def damage_energy_density(state):
     grad_alpha = ufl.grad(alpha)
 
     # Compute the damage dissipation density
-    damage_density = _w1 * w(alpha) + \
-        _w1 * _ell**2 / 2. * ufl.dot(grad_alpha, grad_alpha)
+    damage_density = _w1 * w(alpha) + _w1 * _ell**2 / 2.0 * ufl.dot(
+        grad_alpha, grad_alpha
+    )
 
[…]
