+++ test/test_extend.py
@@ -41,13 +41,13 @@ def test_extend_vector():
     maps = [(V.dofmap.index_map, V.dofmap.index_map_bs) for V in [V_u, V_alpha]]
 
     if getLevelName(_logger.getEffectiveLevel()) == "DEBUG":
-        _logger.debug(f"x")
+        _logger.debug("x")
         x.view()
 
-        _logger.debug(f"vr")
+        _logger.debug("vr")
         vr.view()
 
-    _logger.info(f"Setting up dofs for extension")
+    _logger.info("Setting up dofs for extension")
     _logger.debug(
         f'{__log_incipit} The "good" dofs: constraints.bglobal_dofs_vec_stacked {constraints.bglobal_dofs_vec_stacked}'
     )
@@ -58,11 +58,11 @@ def test_extend_vector():
     _logger.debug(f"{__log_incipit} Local data of the x: {x.array}")
 
     if getLevelName(_logger.getEffectiveLevel()) == "DEBUG":
-        _logger.debug(f"x")
+        _logger.debug("x")
         x.view()
 
     x_u, x_alpha = get_local_vectors(x, maps)
 
-    _logger.info(f"The local vectors")
+    _logger.info("The local vectors")
     _logger.debug(f"{__log_incipit} Local data of the subvector x_u: {x_u}")
     _logger.debug(f"{__log_incipit} Local data of the subvector x_alpha: {x_alpha}")
