+++ src/irrevolutions/practice/parametric-traction-bar-s.py
@@ -1,5 +1,7 @@
 import subprocess
+
 import numpy as np
+
 # List of parameters
 
 # parameters_list = np.logspace(-8, -1, 8)
