+++ src/irrevolutions/meshes/__init__.py
@@ -8,10 +8,7 @@
 
 from functools import wraps
 
-
 # from dolfinx.mesh import create_meshtags, create_mesh
-
-
 from gmsh import model
 
 
