### 2024-08-16 5778265
paper fixes, joss rev jhale
