### 2024-08-16 kumiori/irrevolutions#69 @kumiori
Thank you for pointing out,
we added a concise explanation:

`Irrevolutions` is a computational stability-analysis toolkit designed to solve nonlinear and nonconvex evolutionary problems using advanced numerical methods. It provides efficient algorithms for computing solutions constrained minimum problems with application to irreversible evolutions (hence its name). In particular, this framework is relevant in the context of fracture and damage mechanics.
