### 2024-08-16 kumiori/irrevolutions#62 @kumiori
Thank you for pointing out and providing the snippet.
We have implemented the changes, all the collected tests passed.
