### 2024-08-16 38d0b0a
Update paper.bib
fixed capitalization
