+++ paper/paper.bib
@@ -1,12 +1,222 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% http://bibdesk.sourceforge.net/
 
-%% Created for pierluigi at 2024-07-30 08:23:53 +0900 
+%% Created for pierluigi at 2024-08-16 14:54:57 +0900 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
 
 
+
+@article{SICSIC,
+	abstract = {This paper studies the initiation of cracks in the thermal shock problem through the variational analysis of the quasi-static evolution of a gradient damage model. We consider a two-dimensional semi-infinite slab with an imposed temperature drop on its free surface. The damage model is formulated in the framework of the variational theory of rate-independent processes based on the principles of irreversibility, stability and energy balance. In the case of a sufficiently severe shock, we show that damage immediately occurs and that its evolution follows first a fundamental branch without localization. Then it bifurcates into another branch in which damage localization will take place finally to generate cracks. The determination of the time and mode of that bifurcation allows us to explain the periodic distribution of the so-initiated cracks and to calculate the crack spacing in terms of the material and loading parameters. Numerical investigations complete and quantify the analytical results.},
+	author = {Paul Sicsic and Jean-Jacques Marigo and Corrado Maurini},
+	date-added = {2024-08-02 17:20:53 +0900},
+	date-modified = {2024-08-02 17:21:17 +0900},
+	doi = {10.1016/j.jmps.2013.09.003},
+	journal = {Journal of the Mechanics and Physics of Solids},
+	keywords = {Damage mechanics, Gradient damage model, Thermal shock, Variational methods, Energy balance, Stability and bifurcation, Rayleigh ratio},
+	pages = {256-284},
+	title = {Initiation of a periodic array of cracks in the thermal shock problem: A gradient damage modeling},
+	volume = {63},
+	year = {2014},
+	bdsk-url-1 = {https://www.sciencedirect.com/science/article/pii/S0022509613001786},
+	bdsk-url-2 = {https://doi.org/10.1016/j.jmps.2013.09.003}}
+
+@article{bazant,
+	abstract = {We consider a wide class of gradient damage models which are characterized by two constitutive functions after a normalization of the scalar damage parameter. The evolution problem is formulated following a variational approach based on the principles of irreversibility, stability and energy balance. Applied to a monotonically increasing traction test of a one-dimensional bar, we consider the homogeneous response where both the strain and the damage fields are uniform in space. In the case of a softening behavior, we show that the homogeneous state of the bar at a given time is stable provided that the length of the bar is less than a state dependent critical value and unstable otherwise. However, we also show that bifurcations can appear even if the homogeneous state is stable. All these results are obtained in a closed form. Finally, we propose a practical method to identify the two constitutive functions. This method is based on the measure of the homogeneous response in a situation where this response is stable without possibility of bifurcation, and on a procedure which gives the opportunity to detect its loss of stability. All the theoretical analyses are illustrated by examples.},
+	author = {Zden\v{e}k P. Ba\v{z}ant},
+	date-added = {2024-08-02 17:17:28 +0900},
+	date-modified = {2024-08-16 14:54:04 +0900},
+	doi = {10.1061/(ASCE)0733-9399(1988)114:12(2013)},
+	journal = {Journal of Engineering Mechanics},
+	title = {{Stable States and Paths of Stmuctures with Plasticity or Damage}},
+	volume = {114},
+	year = {1988},
+	bdsk-url-1 = {https://www.sciencedirect.com/science/article/pii/S002250961100055X},
[…]
