### 2024-08-16 kumiori/irrevolutions#53 @kumiori
Thank you,
we took your suggestions and indicated explicitly the sub-labeling of equation 2, for ease and clarity of reference
and reformatted the definition of the minimiser, for improved clarity. Additionally, we fixed typos.
