### 2024-08-16 kumiori/irrevolutions#52 @pcesana81
capitalization in references has been fixed as recommended.
