+++ CODE-OF-CONDUCT.md
@@ -1,0 +1,42 @@
+# Code of Conduct
+
+## Our Pledge
+
+In the interest of developing an open and welcoming environment, we as contributors and maintainers pledge to making participation in our project and our community a smooth experience for everyone, regardless of age, size, personal quirks, ethnicity, gender identity and expression, level of experience, education, socio-economic status, nationality, personal appearance, religion, sex inclinations, or sexual identity and orientation.
+
+## Our Standards
+
+Examples of behavior that contributes to creating a positive environment include:
+
+- Using welcoming and inclusive language
+- Being respectful of differing viewpoints and experiences
+- Gracefully accepting constructive criticism
+- Focusing on what is best for the community
+- Showing empathy towards other community members
+
+Examples of unacceptable behavior by participants include:
+
+- Trolling, insulting/derogatory comments, and personal or political attacks
+- Public or private harassment
+- Publishing others' private information, without explicit permission
+- Other conduct which could reasonably be considered inappropriate in a friendly setting
+
+## Our Responsibilities
+
+Project maintainers are responsible for suggesting best practices of acceptable behavior and are in a position to take appropriate and fair corrective action in response to any instances of unacceptable behavior.
+
+Project maintainers may claim the right and responsibility to remove, edit, or reject comments, commits, code, wiki edits, issues, and other contributions that are not aligned to this Code of Conduct, or to ban temporarily or permanently any contributor for other behaviors that they deem inappropriate, threatening, offensive, or harmful.
+
+## Scope
+
+This Code of Conduct applies both within project spaces and in public spaces when an individual is representing the project or its community. Examples of representing a project or community include using an official project e-mail address, public conferences and scientific gatherings, posting via an official social media account, or acting as an appointed representative at an online or offline event. Representation of a project may be further defined and clarified by project maintainers.
+
+## Enforcement
+
+Instances of abusive, harassing, or otherwise unacceptable behavior may be reported by contacting the project team or in the [Issues page](https://github.com/kumiori/irrevolutions/issues). All complaints will be reviewed and investigated and will result in a response that is deemed necessary and appropriate to the circumstances. The project team maintains confidentiality with regard to the details of an incident. 
+
+Project maintainers are bound to the Code of Conduct as well, failing to honour its best practices may lead to temporary or permanent repercussions as determined by other members of the project's leadership.
+
+## Attribution
+
+This Code of Conduct is adapted from the [Contributor Covenant](https://www.contributor-covenant.org), version 1.4, available at [https://www.contributor-covenant.org/version/1/4/code-of-conduct.html](https://www.contributor-covenant.org/version/1/4/code-of-conduct.html).
