+++ CONTRIBUTING.md
@@ -1,0 +1,34 @@
+# Contributing to Our Project
+We welcome contributions from everyone in the community
+
+### Reporting bugs
+
+If you find a bug in Irrevolutions, please report it on the [GitHub issue
+tracker](https://github.com/fenics/dolfinx/issues/new?labels=bug).
+
+
+### Suggesting enhancements
+
+If you want to suggest a new feature or an improvement of a current
+feature, you can submit this on the [issue
+tracker](https://github.com/fenics/dolfinx/issues).
+
+
+### Submitting a pull request
+
+To contribute code DOLFINx, create a pull request. If you want to
+contribute, but are unsure where to start, have a look at the [Discussions page](https://github.com/kumiori/irrevolutions/discussions).
+For substantial changes/contributions, please start with an issue or
+start a new Discussion thread.
+
+On opening a pull request, unit tests will run on GitHub CI. You can
+click on these in the pull request to see where (if anywhere) the tests
+are failing.
+
+
+### Code of conduct
+
+We suggest all our contributors to read the [code of
+conduct](CODE-OF-CONDUCT.md).
+
+
