+++ CODE-OF-CONDUCT.md
@@ -31,7 +31,7 @@ Project maintainers may claim the right and responsibility to remove, edit, or r
 
 This Code of Conduct applies both within project spaces and in public spaces when an individual is representing the project or its community. Examples of representing a project or community include using an official project e-mail address, public conferences and scientific gatherings, posting via an official social media account, or acting as an appointed representative at an online or offline event. Representation of a project may be further defined and clarified by project maintainers.
 
-## Enforcement
+## Reporting
 
 Instances of abusive, harassing, or otherwise unacceptable behavior may be reported by contacting the project team or in the [Issues page](https://github.com/kumiori/irrevolutions/issues). All complaints will be reviewed and investigated and will result in a response that is deemed necessary and appropriate to the circumstances. The project team maintains confidentiality with regard to the details of an incident. 
 
