### 2024-09-03 c9579c4
sorting and removing unused vars
