+++ src/irrevolutions/solvers/slepcblockproblem.py
@@ -4,7 +4,6 @@ import typing
 # from yaml.tokens import BlockSequenceStartToken
 import dolfinx
 import ufl
-
 # plot_matrix
 from petsc4py import PETSc
 from slepc4py import SLEPc
