+++ src/irrevolutions/meshes/extended_pacman.py
@@ -7,24 +7,12 @@ from mpi4py import MPI
 sys.path.append("../")
 from pathlib import Path
 
-from meshes import (
-    _addCircleArc as addCircleArc,
-)
-from meshes import (
-    _addCurveLoop as addCurveLoop,
-)
-from meshes import (
-    _addLine as addLine,
-)
-from meshes import (
-    _addPhysicalSurface as _addPhysicalSurface,
-)
-from meshes import (
-    _addPlaneSurface as _addPlaneSurface,
-)
-from meshes import (
-    _addPoint as addPoint,
-)
+from meshes import _addCircleArc as addCircleArc
+from meshes import _addCurveLoop as addCurveLoop
+from meshes import _addLine as addLine
+from meshes import _addPhysicalSurface as _addPhysicalSurface
+from meshes import _addPlaneSurface as _addPlaneSurface
+from meshes import _addPoint as addPoint
 
 
 def mesh_extended_pacman(
@@ -128,7 +116,6 @@ if __name__ == "__main__":
     import sys
 
     import yaml
-
     # , merge_meshtags, locate_dofs_topological
     from mpi4py import MPI
 
