+++ src/irrevolutions/practice/traction-AT1_first_order.py
@@ -14,16 +14,8 @@ import petsc4py
 import ufl
 import yaml
 from dolfinx.common import list_timings
-from dolfinx.fem import (
-    Constant,
-    Function,
-    FunctionSpace,
-    assemble_scalar,
-    dirichletbc,
-    form,
-    locate_dofs_geometrical,
-    set_bc,
-)
+from dolfinx.fem import (Constant, Function, FunctionSpace, assemble_scalar,
+                         dirichletbc, form, locate_dofs_geometrical, set_bc)
 from dolfinx.io import XDMFFile, gmshio
 from mpi4py import MPI
 from petsc4py import PETSc
@@ -370,6 +362,7 @@ import sys
 
 import pyvista
 from pyvista.utilities import xvfb
+
 from utils.viz import plot_scalar, plot_vector
 
 #
