+++ src/irrevolutions/meshes/primitives.py
@@ -366,6 +366,7 @@ if __name__ == "__main__":
 
     import dolfinx.plot
     from mesh import gmsh_to_dolfin
+
     # , merge_meshtags, locate_dofs_topological
     from mpi4py import MPI
     from xdmf import XDMFFile
