+++ src/irrevolutions/practice/traction-ATJJ.py
@@ -14,8 +14,16 @@ import petsc4py
 import ufl
 import yaml
 from dolfinx.common import list_timings
-from dolfinx.fem import (Constant, Function, FunctionSpace, assemble_scalar,
-                         dirichletbc, form, locate_dofs_geometrical, set_bc)
+from dolfinx.fem import (
+    Constant,
+    Function,
+    FunctionSpace,
+    assemble_scalar,
+    dirichletbc,
+    form,
+    locate_dofs_geometrical,
+    set_bc,
+)
 from dolfinx.io import XDMFFile, gmshio
 from mpi4py import MPI
 from petsc4py import PETSc
@@ -24,9 +32,10 @@ sys.path.append("../")
 
 from algorithms.am import HybridSolver
 from algorithms.so import BifurcationSolver, StabilitySolver
-from irrevolutions.utils import ColorPrint
 from meshes.primitives import mesh_bar_gmshapi
 from models import DamageElasticityModel
+
+from irrevolutions.utils import ColorPrint
 
 logging.getLogger().setLevel(logging.ERROR)
 
@@ -460,7 +469,6 @@ def traction_with_parameters(parameters, slug=""):
 
     import pyvista
     from pyvista.utilities import xvfb
-
     from utils.viz import plot_scalar, plot_vector
 
     #
@@ -515,7 +523,6 @@ def _plot_bif_spectrum_profile(
     """docstring for _plot_bif_spectrum_profile"""
 
     import matplotlib.pyplot as plt
-
     from utils.viz import plot_profile
 
     # __import__('pdb').set_trace()
@@ -590,7 +597,6 @@ def _plot_bif_spectrum_profile_fullvec(
     """docstring for _plot_bif_spectrum_profile"""
 
     import matplotlib.pyplot as plt
-
     from utils.viz import plot_profile
 
     # fields = data["perturbations_beta"]
@@ -670,7 +676,6 @@ def _plot_perturbations_profile(
     fields, parameters, prefix, plotter=None, label="", idx="", aux=None
 ):
     import matplotlib.pyplot as plt
-
     from utils.viz import plot_profile
 
     u = fields[0]
@@ -740,7 +745,6 @@ def _plot_perturbations_profile(
     # _plt.savefig(f"{prefix}/test_profile{idx}.png")
 
     return plotter, _plt
-
 
 
 def param_ell():
