+++ src/irrevolutions/meshes/extended_pacman.py
@@ -116,6 +116,7 @@ if __name__ == "__main__":
     import sys
 
     import yaml
+
     # , merge_meshtags, locate_dofs_topological
     from mpi4py import MPI
 

[…]
