+++ src/irrevolutions/practice/traction-cone.py
@@ -14,16 +14,8 @@ import petsc4py
 import ufl
 import yaml
 from dolfinx.common import list_timings
-from dolfinx.fem import (
-    Constant,
-    Function,
-    FunctionSpace,
-    assemble_scalar,
-    dirichletbc,
-    form,
-    locate_dofs_geometrical,
-    set_bc,
-)
+from dolfinx.fem import (Constant, Function, FunctionSpace, assemble_scalar,
+                         dirichletbc, form, locate_dofs_geometrical, set_bc)
 from dolfinx.io import XDMFFile, gmshio
 from mpi4py import MPI
 from petsc4py import PETSc
@@ -360,7 +352,8 @@ def traction_with_parameters(parameters, slug=""):
 
     # Viz
 
-    from utils.plots import plot_AMit_load, plot_energies, plot_force_displacement
+    from utils.plots import (plot_AMit_load, plot_energies,
+                             plot_force_displacement)
 
     if comm.rank == 0:
         plot_energies(history_data, file=f"{prefix}/{_nameExp}_energies.pdf")
