+++ test/test_linsearch.py
@@ -17,16 +17,8 @@ import pyvista
 import ufl
 import yaml
 from dolfinx.common import list_timings
-from dolfinx.fem import (
-    Constant,
-    Function,
-    FunctionSpace,
-    assemble_scalar,
-    dirichletbc,
-    form,
-    locate_dofs_geometrical,
-    set_bc,
-)
+from dolfinx.fem import (Constant, Function, FunctionSpace, assemble_scalar,
+                         dirichletbc, form, locate_dofs_geometrical, set_bc)
 from dolfinx.io import XDMFFile, gmshio
 from mpi4py import MPI
 from petsc4py import PETSc
@@ -38,7 +30,8 @@ from irrevolutions.algorithms.so import BifurcationSolver, StabilitySolver
 from irrevolutions.meshes.primitives import mesh_bar_gmshapi
 from irrevolutions.models import DamageElasticityModel as Brittle
 from irrevolutions.solvers.function import vec_to_functions
-from irrevolutions.utils import ColorPrint, _write_history_data, history_data, norm_H1
+from irrevolutions.utils import (ColorPrint, _write_history_data, history_data,
+                                 norm_H1)
 from irrevolutions.utils.plots import plot_energies
 from irrevolutions.utils.viz import plot_profile, plot_scalar, plot_vector
 
