+++ contributed/NOTCH/Notch_problem.py
@@ -1,571 +1,0 @@
-# Numpy -> numerical library for Python. We'll use it for all array operations.
-# It's written in C and it's faster (than traditional Python)
-import logging
-# Yaml (Yet another markup language) -> We'll use it to pass, read and structure
-# light text data in .yml files.
-# Json -> Another form to work with data. It comes from JavaScript. Similar functions
-# that Yaml. Used speacily with API request, when we need data "fetch".
-# Communication with the machine:
-# Sys -> allows to acess the system and launch commandes.
-# Os - > allows to acess the operation system.
-import sys
-
-import dolfinx
-import dolfinx.io
-import dolfinx.plot
-import gmsh
-import matplotlib.pyplot as plt
-import meshes
-import numpy as np
-import pyvista
-import ufl
-from algorithms import am
-from dolfinx.fem import (assemble_scalar, dirichletbc, locate_dofs_geometrical,
-                         set_bc)
-from models import DamageElasticityModel as Brittle
-from petsc4py import PETSc
-from pyvista.utilities import xvfb
-from utils.viz import plot_mesh, plot_scalar, plot_vector
-
-# -> this serves to add a path to the code search for things
-sys.path.append("../")
-
-# pdb -> usefull for debugging, it can stop a code operation and allows to read
-# variables and do calculations
-# pdb.set_trace() #-> point of stop for debugging
-
-# Mpi4py -> Interface that allows parallel interoperability. MPI stands for' Message
-# Passager Interface' and will be used to communicate computer nodes when lauching code
-# in a parallel way
-# Petcs4py -> we use this library to handle with the data. Given acesses
-# to solvers
-
-# Dolfinx
-
-logging.basicConfig(level=logging.INFO)
-
-
-# UFL (Unified Format Language) -> we'll be used to represent abstract way to
-# represent the language in a quadratic form
-
-# XDMFF -> format used for the output binary data
-
-# Install 'gmsh' library -> we'll be used for the mesh.
-#!{sys.executable}: to use the current kernel to make the installation
-
-
-# meshes
-
-# visualisation
-
-# Parameters
-
-parameters = {
-    # In case of evolution (nonlinear) problems, it's necessary to define a max
-    # and a min. For the elastic solution, just one value in needed.
-    "loading": {
-        "type": "ID",  # ID -> Imposed Displacement | IF -> Imposed Force
-        "min": 0,
-        "max": 1.5,
-        "steps": 20,
-    },
-    "geometry": {"geom_type": "bar", "Lx": 1.0, "Ly": 0.01},
-    "model": {
-        "E": 1.0,
-        "nu": 0.3,
-        "mu": 0,  # don't change it -> calculated later
-        "lmbda": 0,  # don't change it -> calculated later
-        "w1": 1.0,
-        "ell": 0.01,
-        "k_res": 1.0e-8,
-    },
-    "solvers": {
-        "elasticity": {
-            "snes": {
-                "snes_type": "newtontr",
-                "snes_stol": 1e-8,
-                "snes_atol": 1e-8,
-                "snes_rtol": 1e-8,
-                "snes_max_it": 100,
-                "snes_monitor": "",
-                "ksp_type": "preonly",
-                "pc_type": "lu",
-                "pc_factor_mat_solver_type": "mumps",
[…]
