+++ contributed/solveModel/solveEP2.py
@@ -1,276 +1,0 @@
-# library include
-
-
-import logging
-import sys
-
-import dolfinx
-import dolfinx.io
-import dolfinx.plot
-import matplotlib.pyplot as plt
-import meshes
-import numpy as np
-import pyvista
-import ufl
-import yaml
-from dolfinx.fem import Function, dirichletbc
-from meshes import primitives
-from petsc4py import PETSc
-from pyvista.utilities import xvfb
-from solvers import SNESSolver
-from utils.viz import plot_mesh, plot_scalar, plot_vector
-
-logging.basicConfig(level=logging.INFO)
-
-
-sys.path.append("./")
-
-# meshes
-
-# visualisation
-
-
-def plot_vector(u, plotter, subplot=None):
-    if subplot:
-        plotter.subplot(subplot[0], subplot[1])
-    V = u.function_space
-    mesh = V.mesh
-    topology, cell_types = dolfinx.plot.create_vtk_topology(mesh, mesh.topology.dim)
-    num_dofs_local = u.function_space.dofmap.index_map.size_local
-    geometry = u.function_space.tabulate_dof_coordinates()[:num_dofs_local]
-    values = np.zeros((V.dofmap.index_map.size_local, 3), dtype=np.float64)
-    values[:, : mesh.geometry.dim] = u.vector.array.real.reshape(
-        V.dofmap.index_map.size_local, V.dofmap.index_map_bs
-    )
-    grid = pyvista.UnstructuredGrid(topology, cell_types, geometry)
-    grid["vectors"] = values
-    grid.set_active_vectors("vectors")
-    # geom = pyvista.Arrow()
-    # glyphs = grid.glyph(orient="vectors", factor=1, geom=geom)
-    glyphs = grid.glyph(orient="vectors", factor=1.0)
-    plotter.add_mesh(glyphs)
-    plotter.add_mesh(
-        grid, show_edges=True, color="black", style="wireframe", opacity=0.3
-    )
-    plotter.view_xy()
-    return plotter
-
-
-def plot_scalar(alpha, plotter, subplot=None, lineproperties={}):
-    if subplot:
-        plotter.subplot(subplot[0], subplot[1])
-    V = alpha.function_space
-    mesh = V.mesh
-    topology, cell_types, _ = dolfinx.plot.create_vtk_mesh(mesh, mesh.topology.dim)
-    grid = pyvista.UnstructuredGrid(topology, cell_types, mesh.geometry.x)
-
-    plotter.subplot(0, 0)
-    grid.point_data["alpha"] = alpha.compute_point_values().real
-    grid.set_active_scalars("alpha")
-    plotter.add_mesh(grid, **lineproperties)
-    plotter.view_xy()
-    return plotter
-
-
-# Parameters
-
-
-parameters = {
-    "loading": {"min": 0, "max": 1},
-    "geometry": {"geom_type": "bar", "Lx": 5.0, "Ly": 15},
-    "model": {"mu": 1.0, "lmbda": 0.0},
-    "solvers": {
-        "snes": {
-            "snes_type": "newtontr",
-            "snes_stol": 1e-8,
-            "snes_atol": 1e-8,
-            "snes_rtol": 1e-8,
-            "snes_max_it": 100,
-            "snes_monitor": "",
-            "ksp_type": "preonly",
-            "pc_type": "lu",
[…]
