+++ src/irrevolutions/practice/traction-cone.py
@@ -1,4 +1,9 @@
 #!/usr/bin/env python3
+from irrevolutions.utils import ColorPrint
+from models import DamageElasticityModel as Brittle
+from meshes.primitives import mesh_bar_gmshapi
+from algorithms.so import BifurcationSolver, StabilitySolver
+from algorithms.am import HybridSolver
 import json
 import logging
 import os
@@ -21,12 +26,7 @@ from mpi4py import MPI
 from petsc4py import PETSc
 
 sys.path.append("../")
-from algorithms.am import HybridSolver
-from algorithms.so import BifurcationSolver, StabilitySolver
-from meshes.primitives import mesh_bar_gmshapi
-from models import DamageElasticityModel as Brittle
 
-from irrevolutions.utils import ColorPrint
 
 logging.getLogger().setLevel(logging.ERROR)
 
