+++ src/irrevolutions/practice/pacman_hybrid.py
@@ -1,3 +1,19 @@
+from irrevolutions.utils import ColorPrint, set_vector_to_constant
+from utils.viz import plot_mesh, plot_scalar, plot_vector
+from utils.lib import _local_notch_asymptotic
+from solvers.function import functions_to_vec
+from petsc4py import PETSc
+from mpi4py import MPI
+from models import DamageElasticityModel as Brittle
+from meshes.pacman import mesh_pacman
+from algorithms.am import HybridSolver
+import ufl
+import petsc4py
+import numpy as np
+import dolfinx
+from datetime import date
+import logging
+import json
 import os
 import sys
 from pathlib import Path
@@ -13,24 +29,8 @@ from dolfinx.mesh import locate_entities_boundary
 from pyvista.utilities import xvfb
 
 sys.path.append("../")
-import json
-import logging
-from datetime import date
 
-import dolfinx
-import numpy as np
-import petsc4py
-import ufl
-from algorithms.am import HybridSolver
-from meshes.pacman import mesh_pacman
-from models import DamageElasticityModel as Brittle
-from mpi4py import MPI
-from petsc4py import PETSc
-from solvers.function import functions_to_vec
-from utils.lib import _local_notch_asymptotic
-from utils.viz import plot_mesh, plot_scalar, plot_vector
 
-from irrevolutions.utils import ColorPrint, set_vector_to_constant
 
 logging.basicConfig(level=logging.INFO)
 
