+++ src/irrevolutions/solvers/__init__.py
@@ -1,3 +1,6 @@
+from dolfinx.fem.petsc import (apply_lifting, assemble_matrix, assemble_vector,
+                               create_matrix, create_vector, set_bc)
+from dolfinx.cpp.log import LogLevel, log
 import sys
 
 import dolfinx
@@ -8,10 +11,7 @@ from petsc4py import PETSc
 
 petsc4py.init(sys.argv)
 
-from dolfinx.cpp.log import LogLevel, log
 # from damage.utils import ColorPrint
-from dolfinx.fem.petsc import (apply_lifting, assemble_matrix, assemble_vector,
-                               create_matrix, create_vector, set_bc)
 
 # import pdb;
 # pdb.set_trace()
