### 2024-09-03 24bca37
regenerate image with JOSS_paper_figures & minor
