+++ test/test_spa.py
@@ -167,7 +167,7 @@ def test_spa():
 
     V_u = dolfinx.fem.FunctionSpace(mesh, element_u)
     V_alpha = dolfinx.fem.FunctionSpace(mesh, element_alpha)
-    u = dolfinx.fem.Function(V_u, name="Displacement")
+    # u = dolfinx.fem.Function(V_u, name="Displacement")
     alpha = dolfinx.fem.Function(V_alpha, name="Damage")
     ufl.Measure("dx", alpha.function_space.mesh)
 
@@ -213,7 +213,7 @@ def test_spa():
         f"lambda_0 = {lmbda_t:.4e}, residual norm = {y.norm(): .4e}, error = {errors[-1]: .4e}"
     )
 
-    assert np.isclose(lmbda_t, -0.044659195907104675, atol=1e-4) == True
+    assert np.isclose(lmbda_t, -0.044659195907104675, atol=1e-4)
 
 
 if __name__ == "__main__":
