+++ src/irrevolutions/utils/eigenspace.py
@@ -187,3 +187,32 @@ def book_of_the_numbers(scale_b=3, scale_c=3):
             break
 
     return {"a": a, "b": b, "c": c}
+
+
+def l2_norm(components):
+    """
+    Compute the L2 norm of a vector field defined by components on their respective domains.
+
+    Parameters:
+        components (list): List of tuples (x, f) where x is the mesh coordinates and f is the field values.
+
+    Returns:
+        float: L2 norm of the vector field.
+    """
+    norms = []
+
+    for x, f in components:
+        # Compute the square of the field values
+        squared_values = f**2
+
+        # Compute the integral using the trapezoidal rule
+        integral = np.trapz(squared_values, x)
+
+        # Take the square root to get the L2 norm
+        component_norm = np.sqrt(integral)
+        norms.append(component_norm)
+
+    # Compute the vector norm by summing the squared norms and taking the square root
+    vector_norm = np.sqrt(np.sum(np.array(norms)**2))
+
+    return vector_norm
