### 2024-09-03 kumiori/irrevolutions#68 @kumiori
Thank you for pointing out the missing element. We added a CONTRIBUTING.md and a CODE-OF-CONDUCT.md
