### 2024-09-03 kumiori/irrevolutions#65 @kumiori
You can run the JOSS_paper_figures_rayleigh_benchmark.ipynb file, which now correctly includes the package's library.
