### 2024-09-03 kumiori/irrevolutions#67 @kumiori
We have added one line in the README.md

This project contains code from the DOLFINy project (https://github.com/fenics-dolfiny/dolfiny), which is licensed under the LGPLv3 license. We acknowledge and thank the DOLFINy contributors for their work.

Should we include in all files with code from DOLFINy the license header?
