### 2024-09-04 #6897 @jorgensd
@kumiori  I think you have changed the access to your repo without meaning to.
One can no longer comment on issues or open new issues:
![image](https://github.com/user-attachments/assets/6a828ae0-126d-4496-b948-2bb7f5a929f1)
![image](https://github.com/user-attachments/assets/ab14f45f-1a67-4e70-92ab-9b67005fbb64)
