### 2024-09-04 #6897 @jorgensd
I also cannot open Pull Request, for instance to fix your contributing file:
![image](https://github.com/user-attachments/assets/60aa53c0-dd2b-4535-8f8c-08a0214d2009)
