+++ test/test_sample_data.py
@@ -8,6 +8,7 @@ from dolfinx.cpp.la.petsc import get_local_vectors, scatter_local_vectors
 from irrevolutions.utils import _logger
 from mpi4py import MPI
 from petsc4py import PETSc
+import basix.ufl
 
 sys.path.append("../")
 
@@ -22,11 +23,11 @@ def init_data(N, positive=True):
     comm = MPI.COMM_WORLD
     comm.Get_rank()
 
-    element_u = ufl.FiniteElement("Lagrange", mesh.ufl_cell(), degree=1)
-    element_alpha = ufl.FiniteElement("Lagrange", mesh.ufl_cell(), degree=1)
+    element_u = basix.ufl.element("Lagrange", mesh.basix_cell(), degree=1)
+    element_alpha = basix.ufl.element("Lagrange", mesh.basix_cell(), degree=1)
 
-    V_u = dolfinx.fem.FunctionSpace(mesh, element_u)
-    V_alpha = dolfinx.fem.FunctionSpace(mesh, element_alpha)
+    V_u = dolfinx.fem.functionspace(mesh, element_u)
+    V_alpha = dolfinx.fem.functionspace(mesh, element_alpha)
     u = dolfinx.fem.Function(V_u, name="Displacement")
     alpha = dolfinx.fem.Function(V_alpha, name="Damage")
     dx = ufl.Measure("dx", alpha.function_space.mesh)
