+++ src/irrevolutions/practice/default.py
@@ -43,6 +43,8 @@ from dolfinx.fem import (
     locate_dofs_geometrical,
     set_bc,
 )
+import basix.ufl
+
 from dolfinx.io import XDMFFile, gmshio
 from mpi4py import MPI
 from petsc4py import PETSc
@@ -147,10 +149,10 @@ def create_function_space(mesh):
         dolfinx.FunctionSpace: Function space for displacement.
         dolfinx.FunctionSpace: Function space for damage.
     """
-    element_u = ufl.VectorElement("Lagrange", mesh.ufl_cell(), degree=1, dim=2)
+    element_u = basix.ufl.element("Lagrange", mesh.basix_cell(), degree=1, shape=(2,))
     V_u = FunctionSpace(mesh, element_u)
 
-    element_alpha = ufl.FiniteElement("Lagrange", mesh.ufl_cell(), degree=1)
+    element_alpha = basix.ufl.element("Lagrange", mesh.basix_cell(), degree=1)
     V_alpha = FunctionSpace(mesh, element_alpha)
 
     return V_u, V_alpha
