+++ src/irrevolutions/practice/traction-cone.py
@@ -153,7 +153,7 @@ def traction_with_parameters(parameters, slug=""):
     alpha_ub.interpolate(lambda x: np.ones_like(x[0]))
 
     for f in [zero_u, zero_alpha, u_, alpha_lb, alpha_ub]:
-        f.vector.ghostUpdate(
+        f.x.petsc_vec.ghostUpdate(
             addv=PETSc.InsertMode.INSERT, mode=PETSc.ScatterMode.FORWARD
         )
 
@@ -168,8 +168,8 @@ def traction_with_parameters(parameters, slug=""):
     #     dolfinx.fem.dirichletbc(zeroKalpha, dofs_alpha_right),
     # ]
 
-    set_bc(alpha_ub.vector, bcs_alpha)
-    alpha_ub.vector.ghostUpdate(
+    set_bc(alpha_ub.x.petsc_vec, bcs_alpha)
+    alpha_ub.x.petsc_vec.ghostUpdate(
         addv=PETSc.InsertMode.INSERT, mode=PETSc.ScatterMode.FORWARD
     )
 
@@ -244,13 +244,13 @@ def traction_with_parameters(parameters, slug=""):
     for i_t, t in enumerate(loads):
         # for i_t, t in enumerate([0., .99, 1.0, 1.01]):
         u_.interpolate(lambda x: (t * np.ones_like(x[0]), np.zeros_like(x[1])))
-        u_.vector.ghostUpdate(
+        u_.x.petsc_vec.ghostUpdate(
             addv=PETSc.InsertMode.INSERT, mode=PETSc.ScatterMode.FORWARD
         )
 
         # update the lower bound
-        alpha.vector.copy(alpha_lb.vector)
-        alpha_lb.vector.ghostUpdate(
+        alpha.x.petsc_vec.copy(alpha_lb.x.petsc_vec)
+        alpha_lb.x.petsc_vec.ghostUpdate(
             addv=PETSc.InsertMode.INSERT, mode=PETSc.ScatterMode.FORWARD
         )
 
@@ -272,19 +272,19 @@ def traction_with_parameters(parameters, slug=""):
         hybrid.solve(alpha_lb)
 
         # compute the rate
-        alpha.vector.copy(alphadot.vector)
-        alphadot.vector.axpy(-1, alpha_lb.vector)
-        alphadot.vector.ghostUpdate(
+        alpha.x.petsc_vec.copy(alphadot.x.petsc_vec)
+        alphadot.x.petsc_vec.axpy(-1, alpha_lb.x.petsc_vec)
+        alphadot.x.petsc_vec.ghostUpdate(
             addv=PETSc.InsertMode.INSERT, mode=PETSc.ScatterMode.FORWARD
         )
 
         rate_12_norm = hybrid.scaled_rate_norm(alpha, parameters)
         urate_12_norm = hybrid.unscaled_rate_norm(alpha)
 
-        logging.critical(f"alpha vector norm: {alpha.vector.norm()}")
-        logging.critical(f"alpha lb norm: {alpha_lb.vector.norm()}")
-        logging.critical(f"alphadot norm: {alphadot.vector.norm()}")
-        logging.critical(f"vector norms [u, alpha]: {[zi.vector.norm() for zi in z]}")
+        logging.critical(f"alpha vector norm: {alpha.x.petsc_vec.norm()}")
+        logging.critical(f"alpha lb norm: {alpha_lb.x.petsc_vec.norm()}")
+        logging.critical(f"alphadot norm: {alphadot.x.petsc_vec.norm()}")
+        logging.critical(f"vector norms [u, alpha]: {[zi.x.petsc_vec.norm() for zi in z]}")
         logging.critical(f"scaled rate state_12 norm: {rate_12_norm}")
[…]
