### 2024-09-05 1b461d4
Cannot get it running with newtontr
