### 2024-09-10 kumiori/irrevolutions#65 @jorgensd
The benchmark seems to depend on some files from the test directory:
https://github.com/kumiori/irrevolutions/blob/main/playground/nb/JOSS_paper_figures_rayleigh_benchmark.ipynb?short_path=473f03c#L139-L140
What tests in particular should one run first?:)
