### 2024-09-16 kumiori/irrevolutions#65 @jorgensd
I tried running the tests, and subsequently modifying the hardcoded output paths in the notebook, but it fails at:
```bash
ls: cannot access '/root/shared/test/output/rayleigh-benchmark-parametric/MPI-1': No such file or directory
```
and to get that output I need to run `rayleigh-benchmark.ipynb` which as not been updated (https://github.com/jorgensd/irrevolutions/blob/main/playground/nb/rayleigh_benchmark.ipynb)
