### 2024-09-16 kumiori/irrevolutions#78 @jorgensd
Document functions in `irrevolutions`
Ref: https://github.com/openjournals/joss-reviews/issues/6897

Would be good to have some docstrings for the functions in:
https://github.com/kumiori/irrevolutions/blob/main/src/irrevolutions/utils/parametric.py
https://github.com/kumiori/irrevolutions/blob/main/src/irrevolutions/utils/plots.py
https://github.com/kumiori/irrevolutions/blob/main/src/irrevolutions/utils/viz.py
https://github.com/kumiori/irrevolutions/blob/main/src/irrevolutions/solvers/restriction.py#L69-L100
https://github.com/kumiori/irrevolutions/blob/main/src/irrevolutions/solvers/slepcblockproblem.py#L118-L146
https://github.com/kumiori/irrevolutions/blob/main/src/irrevolutions/solvers/slepcblockproblem.py#L274-L364
https://github.com/kumiori/irrevolutions/blob/main/src/irrevolutions/solvers/snesblockproblem.py#L292-L383
https://github.com/kumiori/irrevolutions/blob/main/src/irrevolutions/algorithms/am.py
https://github.com/kumiori/irrevolutions/blob/main/src/irrevolutions/algorithms/ls.py
