### 2024-09-16 #6897 @jorgensd
> @kumiori @jorgensd how are the fixes coming along? @jhale, I see you still have some unchecked boxes on your review, can you advise? Thanks.

I have not been able to reproduce the figures from the paper, as mentioned in: https://github.com/kumiori/irrevolutions/issues/65

Other than that, I think some documentation is needed for some of the functions in the library:
https://github.com/kumiori/irrevolutions/issues/78

Once those are fixed, I am happy with the state of the software.
