### 2024-09-23 caa4ab4
merge docstrings
