+++ src/irrevolutions/utils/parametric.py
@@ -1,61 +1,115 @@
 import hashlib
-
 import yaml
 
-
 def parameters_vs_ell(parameters=None, ell=0.1):
-    if parameters is None:
+    """
+    Update the model parameters for a given value of 'ell'.
+
+    This function modifies the 'ell' parameter. 
+    If no parameters are provided, it loads them from the default file.
+
+    Args:
+        parameters (dict, optional): Dictionary of parameters. 
+                                      If None, load from "../test/parameters.yml".
+        ell (float, optional): The new 'ell' value to set in the parameters. 
+                               Default is 0.1.
+
+    Returns:
+        tuple: A tuple containing the updated parameters dictionary and 
+               a unique hash (signature) based on the updated parameters.
+    """
+    if parameters is None:    
         with open("../test/parameters.yml") as f:
             parameters = yaml.load(f, Loader=yaml.FullLoader)
-
+        
     parameters["model"]["ell"] = ell
-    parameters["geometry"]["ell_lc"] = 3
 
-    signature = hashlib.md5(str(parameters).encode("utf-8")).hexdigest()
+    # Generate a unique signature using MD5 hash based on the updated parameters
+    signature = hashlib.md5(str(parameters).encode('utf-8')).hexdigest()
 
     return parameters, signature
 
 
 def parameters_vs_elle(parameters=None, elle=0.3):
-    # for the thin film model
-    if parameters is None:
+    """
+    Update the model parameters for a given value of 'elle' in the thin film model.
+    ell_e is the elastic lengthscale, depending on the film and the substrate properties.
+    
+    Args:
+        parameters (dict, optional): Dictionary of parameters. 
+                                      If None, load from "../data/thinfilm/parameters.yml".
+        elle (float, optional): The new 'ell_e' value to set in the parameters. 
+                                Default is 0.3.
+
+    Returns:
+        tuple: A tuple containing the updated parameters dictionary and 
+               a unique hash (signature) based on the updated parameters.
+    """
+    if parameters is None:    
         with open("../data/thinfilm/parameters.yml") as f:
             parameters = yaml.load(f, Loader=yaml.FullLoader)
 
     parameters["model"]["ell_e"] = elle
-    parameters["model"]["ell"] = elle / 3
-    parameters["geometry"]["mesh_size_factor"] = 3
-    parameters["geometry"]["lc"] = (
-        parameters["model"]["ell"] / parameters["geometry"]["mesh_size_factor"]
-    )
-    signature = hashlib.md5(str(parameters).encode("utf-8")).hexdigest()
+    # parameters["model"]["ell"] = elle / 3
+
+    # Generate a unique signature using MD5 hash based on the updated parameters
+    signature = hashlib.md5(str(parameters).encode('utf-8')).hexdigest()
 
     return parameters, signature
 
 
 def parameters_vs_SPA_scaling(parameters=None, s=0.01):
-    if parameters is None:
+    """
+    Update the stability scaling parameter for the SPA (Second-Order Analysis).
+
+    This function updates the scaling factor in the cone algorithm for the stability analysis.
+
+    Args:
+        parameters (dict, optional): Dictionary of parameters. 
+                                      If None, load from "../test/parameters.yml".
+        s (float, optional): The new scaling value to set in the parameters. 
+                             Default is 0.01.
+
+    Returns:
+        tuple: A tuple containing the updated parameters dictionary and 
+               a unique hash (signature) based on the updated parameters.
+    """
+    if parameters is None:    
         with open("../test/parameters.yml") as f:
             parameters = yaml.load(f, Loader=yaml.FullLoader)
 
     parameters["stability"]["cone"]["scaling"] = s
 
[…]
