### 2024-09-24 687d787
docstrings in am and ls, removing unused file
