+++ src/irrevolutions/algorithms/ls.py
@@ -1,3 +1,4 @@
+import hashlib
 import logging
 import random
 
@@ -6,30 +7,38 @@ import numpy as np
 from dolfinx.cpp.log import LogLevel, log
 from dolfinx.fem import Function, assemble_scalar, form
 from petsc4py import PETSc
-
 from irrevolutions.utils import norm_H1
 
 # Set up logging configuration
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
-
 comm = mpi4py.MPI.COMM_WORLD
 rank = comm.Get_rank()
 size = comm.Get_size()
 
-# create a class naled linesearch
 
+class LineSearch:
+    """A class to perform line search algorithms for variational problems.
 
-class LineSearch(object):
+    Attributes:
+        energy (ufl.Form): The total energy functional.
+        state (dict): The current state of the system with displacement ('u') and damage ('alpha').
+        parameters (dict): Line search parameters such as method and interpolation order.
+    """
+
     def __init__(self, energy, state, linesearch_parameters={}):
-        super(LineSearch, self).__init__()
-        # self.u_0 = dolfin.Vector(state['u'].vector())
-        # self.alpha_0 = dolfin.Vector(state['alpha'].vector())
+        """
+        Initialize LineSearch with energy functional, state, and optional parameters.
 
+        Parameters:
+            energy (ufl.Form): The total energy functional.
+            state (dict): The current state containing the displacement ('u') and damage ('alpha').
+            linesearch_parameters (dict): Parameters for the line search algorithm.
+        """
+        super(LineSearch, self).__init__()
         self.energy = energy
         self.state = state
-        # initial state
         self.parameters = linesearch_parameters
 
         self.V_u = state["u"].function_space
@@ -40,8 +49,19 @@ class LineSearch(object):
         self.alpha0 = Function(state["alpha"].function_space)
 
     def search(self, state, perturbation, interval, m=2, method="min"):
-        # m = self.parameters["order"]
+        """
+        Perform a line search using a polynomial interpolation method.
 
+        Parameters:
+            state (dict): The current state with displacement ('u') and damage ('alpha').
+            perturbation (dict): Perturbation fields containing the displacement perturbation ('v') and damage perturbation ('beta').
+            interval (tuple): A tuple specifying the admissible search interval (hmin, hmax).
+            m (int): The interpolation order (default is 2).
+            method (str): The method for line search ('min' for minimum, 'random' for random choice).
+
+        Returns:
+            tuple: The optimal step size h_opt, energies along the search direction, the fitted polynomial, and coefficients.
+        """
         v = perturbation["v"]
         beta = perturbation["beta"]
 
@@ -53,29 +73,18 @@ class LineSearch(object):
 
         en_0 = assemble_scalar(form(self.energy))
 
-        # get admissible interval
         hmin, hmax = interval
-
-        # discretise interval for polynomial interpolation at order m
         htest = np.linspace(hmin, hmax, np.int32(m + 1))
         energies_1d = []
         perturbation_norms = []
 
-        # compute energy at discretised points
-
         for h in htest:
-            with state["u"].vector.localForm() as u_local, state[
-                "alpha"
-            ].vector.localForm() as alpha_local, alpha_0.vector.localForm() as alpha0_local, u_0.vector.localForm() as u0_local, v.vector.localForm() as v_local, beta.vector.localForm() as beta_local:
-                u_local.array[:] = u0_local.array[:] + h * v_local.array[:]
[…]
