+++ src/irrevolutions/algorithms/am.py
@@ -1,4 +1,3 @@
-from dolfinx.fem.petsc import assemble_vector, set_bc
 import logging
 
 import dolfinx
@@ -14,14 +13,41 @@ from irrevolutions.solvers.function import functions_to_vec
 from irrevolutions.solvers.snesblockproblem import SNESBlockProblem
 from irrevolutions.utils import (ColorPrint, norm_H1, norm_L2,
                                  set_vector_to_constant)
+from dolfinx.fem.petsc import assemble_vector, set_bc
 
 comm = MPI.COMM_WORLD
 
-
+# Set up basic logging
 logging.basicConfig()
 
-
 class AlternateMinimisation:
+    """
+    First order Alternate Minimisation solver for elasticity and damage fields.
+
+    The solver seeks equilibia (critical states for the total energy) of a system 
+    by alternating between solving for the elasticity and the damage field in an iterative process,
+    exploiting the separate convexity of the total energy.
+    
+    Remark: check the assumptions.
+
+    Parameters:
+    ----------
+    total_energy : ufl.Form
+        Total energy functional of the system to be minimized.
+    state : dict
+        Dictionary containing the current state of the system with keys 'u' (displacement)
+        and 'alpha' (damage).
+    bcs : dict
+        Dictionary of boundary conditions with keys 'bcs_u' for displacement and 'bcs_alpha' 
+        for damage.
+    solver_parameters : dict, optional
+        Dictionary of solver parameters, default is an empty dictionary.
+    bounds : tuple of dolfinx.fem.function.Function, optional
+        Tuple containing lower and upper bounds for the damage field (default: no bounds).
+    monitor : callable, optional
+        Function to monitor progress, if provided.
+    """
+
     def __init__(
         self,
         total_energy,
@@ -31,9 +57,9 @@ class AlternateMinimisation:
         bounds=(dolfinx.fem.function.Function, dolfinx.fem.function.Function),
         monitor=None,
     ):
+        # Initialize state and energy
         self.u = state["u"]
         self.alpha = state["alpha"]
-        # self.bcs  = bcs
         self.alpha_old = Function(self.alpha.function_space)
         self.alpha.vector.copy(self.alpha_old.vector)
         self.alpha.vector.ghostUpdate(
@@ -41,41 +67,30 @@ class AlternateMinimisation:
         )
 
         self.total_energy = total_energy
-
         self.state = state
         self.alpha_lb = bounds[0]
         self.alpha_ub = bounds[1]
-        self.total_energy = total_energy
-        # self.solver_parameters = default_parameters["solvers"]
-        # if solver_parameters:
         self.solver_parameters = solver_parameters
-
         self.monitor = monitor
-
-        # self.dx = ufl.Measure("dx", self.alpha.function_space.mesh)
 
         V_u = self.u.function_space
         V_alpha = self.alpha.function_space
 
+        # Derivatives of the total energy
         energy_u = ufl.derivative(self.total_energy, self.u, ufl.TestFunction(V_u))
         energy_alpha = ufl.derivative(
             self.total_energy, self.alpha, ufl.TestFunction(V_alpha)
         )
         self.F = [energy_u, energy_alpha]
 
+        # SNESSolver for elasticity and damage
         self.elasticity = SNESSolver(
             energy_u,
             self.u,
             bcs.get("bcs_u"),
-            bounds=None,
             petsc_options=self.solver_parameters.get("elasticity").get("snes"),
             prefix=self.solver_parameters.get("elasticity").get("prefix"),
         )
-        # Set near nullspace for the gamg preconditioner for elasticity
-
[…]
