+++ src/irrevolutions/solvers/slepcblockproblem.py
@@ -1,10 +1,8 @@
 import logging
 import typing
 
-# from yaml.tokens import BlockSequenceStartToken
 import dolfinx
 import ufl
-# plot_matrix
 from petsc4py import PETSc
 from slepc4py import SLEPc
 
@@ -22,33 +20,34 @@ class SLEPcBlockProblem:
         B_form=None,
         prefix=None,
     ):
-        """SLEPc problem and solver wrapper.
+        """
+        Initialize the SLEPcBlockProblem for solving a generalised eigenvalue problem.
 
         Wrapper for a generalised eigenvalue problem obtained from UFL residual forms.
-
+        
         Parameters
         ----------
-        F_form
-            Residual forms
-        u
-            Current solution vectors
-        lmbda
-            Eigenvalue function. Residual forms must be linear in lmbda for
-            linear eigenvalue problems.
-        bcs
+        F_form : List
+            Residual forms.
+        u : List
+            Current solution vectors.
+        lmbda : dolfinx.fem.Function
+            Eigenvalue function. Residual forms must be linear in lambda for linear eigenvalue problems.
+        bcs : List, optional
             List of boundary conditions.
-        A_form, optional
-            Override automatically derived A
-        B_form, optional
-            Override automatically derived B
+        A_form : optional
+            Override automatically derived A matrix.
+        B_form : optional
+            Override automatically derived B matrix.
+        prefix : str, optional
+            Prefix for SLEPc options.
 
         Note
         ----
-        In general, eigenvalue problems have form T(lmbda) * x = 0,
-        where T(lmbda) is a matrix-valued function.
-        Linear eigenvalue problems have T(lmbda) = A + lmbda * B, and if B is not identity matrix
-        then this problem is called generalized (linear) eigenvalue problem.
-
+        In general, eigenvalue problems have the form T(lambda) * x = 0,
+        where T(lambda) is a matrix-valued function.
+        Linear eigenvalue problems have T(lambda) = A + lambda * B, and if B is not identity matrix
+        then this problem is called a generalized (linear) eigenvalue problem.
         """
         self.F_form = F_form
         self.u = u
@@ -116,6 +115,11 @@ class SLEPcBlockProblem:
             self.B = dolfinx.fem.create_matrix_block(self.B_form)
 
     def solve(self):
+        """
+        Solve the generalized eigenvalue problem.
+
+        Assemble the matrices A and B, and solve the eigenvalue problem using SLEPc.
+        """
         self.A.zeroEntries()
         dolfinx.fem.assemble_matrix_block(self.A, self.A_form, self.bcs)
         self.A.assemble()
@@ -129,6 +133,23 @@ class SLEPcBlockProblem:
         self.eps.solve()
 
     def getEigenpair(self, i):
+        """
+        Get the eigenvalue and eigenvector pair.
+
+        Parameters
+        ----------
+        i : int
+            Index of the eigenvalue/eigenvector to retrieve.
+
+        Returns
+        -------
+        eigval : float
+            The eigenvalue corresponding to the i-th eigenpair.
+        ur : List[dolfinx.fem.Function]
+            The real part of the eigenvector.
+        ui : List[dolfinx.fem.Function]
+            The imaginary part of the eigenvector.
+        """
         xr, xi = self.A.getVecs()
         eigval = self.eps.getEigenpair(i, xr, xi)
 
@@ -138,13 +159,20 @@ class SLEPcBlockProblem:
         return (eigval, self.ur, self.ui)
 
     def empty_B(self):
+        """
+        Check if the B matrix is empty.
+
+        Returns
+        -------
+        bool
+            True if B matrix is empty, False otherwise.
+        """
[…]
