### 2024-09-24 eadf1f2
pandas poetry requirements
