### 2024-09-24 6c2c93c
documentation configuration
