### 2024-09-24 0e32f9a
minor ruff fixes
