+++ src/irrevolutions/algorithms/ls.py
@@ -1,4 +1,3 @@
-import hashlib
 import logging
 import random
 
