### 2024-09-24 kumiori/irrevolutions#78 @kumiori
We have further the source and enhanced the indicated files with docstrings.
The generated documentation reflects the changes
https://irreversible-variational-solvers.readthedocs.io/en/latest/

(the automatic structure of the rtd references still needs some fixes)
