### 2024-09-24 kumiori/irrevolutions#65 @kumiori
Good point. I have reproduced the figures 1-5 from the paper starting from a newly cloned git repository.
I did run
- test/test_rayleigh.py and
- test/test_rayleigh_parametric.py

to generate the data in test/output/rayleigh-benchmark and rayleigh-benchmark-parametric, respectively.

Then, I ran the updated notebook JOSS_paper_figures_rayleigh_benchmark.ipynb, which runs provided that the variable experiment is initialised to point to the rayleigh-benchmark output (i.e. test/output/MPI-1/<your hash>).

On the other hand, the parametric phase diagram is generated by traversing the rayleigh-benchmark-parametric output folder.

This should be enough to reproduce the figures, will look different from those in the paper because the input data is randomly generated.
