+++ src/irrevolutions/practice/traction-parametric.py
@@ -458,7 +458,7 @@ def param_vs_ell():
     # __import__('pdb').set_trace()
 
     for ell in ell_list:
-        parameters, signature = parameters_vs_ell(parameters= base_parameters, ell=ell)
+        parameters, signature = parameters_vs_ell(parameters=base_parameters, ell=ell)
         _storage = f"output/parametric/traction-bar/vs_ell/{base_signature}/{signature}"
         history_data, performance, state = main(parameters, _storage)
 
@@ -480,7 +480,7 @@ def param_vs_s(base_parameters, base_signature):
 
     for s in s_list:
         parameters, signature = parameters_vs_SPA_scaling(
-            parameters= base_parameters, s=s
+            parameters=base_parameters, s=s
         )
         _storage = f"output/parametric/traction-bar/vs_s/{base_signature}/{signature}"
 
@@ -530,7 +530,7 @@ def param_vs_dry(base_parameters, base_signature):
             yaml.dump(base_parameters, file)
 
     for s in s_list:
-        # parameters, signature = parameters_vs_SPA_scaling(parameters= base_parameters, s=s)
+        # parameters, signature = parameters_vs_SPA_scaling(parameters=base_parameters, s=s)
         parameters = base_parameters
         signature = s
         _storage = f"output/parametric/traction-bar/dry/{signature}"
@@ -590,13 +590,13 @@ if __name__ == "__main__":
 
     if "-s" in sys.argv:
         parameters, signature = parameters_vs_SPA_scaling(
-            parameters= base_parameters, s=np.float(args.s)
+            parameters=base_parameters, s=np.float(args.s)
         )
         _storage = f"output/parametric/traction-bar/vs_s/{args.model}/{base_signature}/{signature}"
 
     elif "-n" in sys.argv:
         parameters, signature = parameters_vs_n_refinement(
-            parameters= base_parameters, r=np.int(args.n)
+            parameters=base_parameters, r=np.int(args.n)
         )
         _storage = f"output/parametric/traction-bar/vs_resolution/{args.model}/{base_signature}/{signature}"
 
