### 2024-09-27 #6897 @majensen
@kumiori have you been able to respond to the suggestions? See https://github.com/openjournals/joss-reviews/issues/6897#issuecomment-2352143470. Thanks
