### 2024-10-02 940c636
Replace remaining vecs
