+++ test/test_1d.py
@@ -120,9 +120,9 @@ class _AlternateMinimisation1D:
                 (solver_alpha_it, solver_alpha_reason) = self.damage.solve()
 
             # Define error function
-            self.alpha.x.petsc_vec.copy(alpha_diff.vector)
-            alpha_diff.vector.axpy(-1, self.alpha_old.vector)
-            alpha_diff.vector.ghostUpdate(
+            self.alpha.x.petsc_vec.copy(alpha_diff.x.petsc_vec)
+            alpha_diff.x.petsc_vec.axpy(-1, self.alpha_old.x.petsc_vec)
+            alpha_diff.x.petsc_vec.ghostUpdate(
                 addv=PETSc.InsertMode.INSERT, mode=PETSc.ScatterMode.FORWARD
             )
 
@@ -135,7 +135,7 @@ class _AlternateMinimisation1D:
                 np.array([comm.allreduce(Fvi.norm(), op=MPI.SUM) for Fvi in Fv]).sum()
             )
 
-            error_alpha_max = alpha_diff.vector.max()[1]
+            error_alpha_max = alpha_diff.x.petsc_vec.max()[1]
             total_energy_int = comm.allreduce(
                 assemble_scalar(form(self.total_energy)), op=MPI.SUM
             )
@@ -143,28 +143,28 @@ class _AlternateMinimisation1D:
             residual_F.ghostUpdate(
                 addv=PETSc.InsertMode.ADD, mode=PETSc.ScatterMode.REVERSE
             )
-            set_bc(residual_F, self.elasticity.bcs, self.u.vector)
+            set_bc(residual_F, self.elasticity.bcs, self.u.x.petsc_vec)
             error_residual_F = ufl.sqrt(residual_F.dot(residual_F))
 
-            self.alpha.x.petsc_vec.copy(self.alpha_old.vector)
-            self.alpha_old.vector.ghostUpdate(
+            self.alpha.x.petsc_vec.copy(self.alpha_old.x.petsc_vec)
+            self.alpha_old.x.petsc_vec.ghostUpdate(
                 addv=PETSc.InsertMode.INSERT, mode=PETSc.ScatterMode.FORWARD
             )
 
             logging.critical(
-                f"AM - Iteration: {iteration:3d}, res F Error: {error_residual_F:3.4e}, alpha_max: {self.alpha.vector.max()[1]:3.4e}"
+                f"AM - Iteration: {iteration:3d}, res F Error: {error_residual_F:3.4e}, alpha_max: {self.alpha.x.petsc_vec.max()[1]:3.4e}"
             )
 
             logging.critical(
-                f"AM - Iteration: {iteration:3d}, H1 Error: {error_alpha_H1:3.4e}, alpha_max: {self.alpha.vector.max()[1]:3.4e}"
+                f"AM - Iteration: {iteration:3d}, H1 Error: {error_alpha_H1:3.4e}, alpha_max: {self.alpha.x.petsc_vec.max()[1]:3.4e}"
             )
 
             logging.critical(
-                f"AM - Iteration: {iteration:3d}, L2 Error: {error_alpha_L2:3.4e}, alpha_max: {self.alpha.vector.max()[1]:3.4e}"
+                f"AM - Iteration: {iteration:3d}, L2 Error: {error_alpha_L2:3.4e}, alpha_max: {self.alpha.x.petsc_vec.max()[1]:3.4e}"
             )
 
             logging.critical(
-                f"AM - Iteration: {iteration:3d}, Linfty Error: {error_alpha_max:3.4e}, alpha_max: {self.alpha.vector.max()[1]:3.4e}"
[…]
