### 2024-10-02 53e9352
Replace scale with alpha
