### 2024-10-05 kumiori/irrevolutions#67 @jhale
Yes, given the similarity the authors of dolfiny should be retained in the files. If you make modifications you can add your own name.
