### 2024-10-07 656a3c7
installation instructions and non-code files handling
