+++ pyproject.toml
@@ -26,3 +26,18 @@ dolfinx_support = []
 
 [tool.poetry.scripts]
 irrevolutions = 'irrevolutions:main'
+
+# Ensure that non-code files (e.g., default_parameters.yml) are included in the package
+[tool.poetry.include]
+irrevolutions = ["src/irrevolutions/models/default_parameters.yml"]
+
+# Additional files to include (e.g., readme, licenses, configs)
+include = [
+    "README.md",
+    "LICENSE",
+    "src/irrevolutions/models/default_parameters.yml"
+]
+
+[tool.poetry.package]
+# Map the source code to the correct directory structure
+package = { "irrevolutions": "src/irrevolutions" }
