+++ .github/workflows/ubuntu.yml
@@ -1,0 +1,49 @@
+name: Test Ubuntu Installation
+
+# This triggers the workflow on any push to the repository
+on: 
+  push:
+    branches:
+      - "**"
+  pull_request:
+
+jobs:
+  test:
+    name: Test Ubuntu/apt Installation and Setup
+    runs-on: ubuntu-latest
+
+    # Steps to run
+    steps:
+      # Checkout the repo content to the runner
+      - name: Checkout repository
+        uses: actions/checkout@v2
+
+      # Install Ubuntu dependencies via APT
+      - name: Install system dependencies via APT
+        run: |
+          echo "Installing required system dependencies via APT"
+          sudo apt-get update
+          sudo apt-get install -y  software-properties-common python3-pip git libgl1-mesa-glx xvfb libglu1 libxcursor1 libxinerama1
+
+      # Add the FEniCS PPA and install fenicsx
+      - name: Install fenicsx
+        run: |
+          echo "Adding FEniCS PPA and installing fenicsx"
+          sudo add-apt-repository ppa:fenics-packages/fenics
+          sudo apt update
+          sudo apt-get install -y fenicsx
+
+      # Install irrevolutions dependencies (if any) and the package itself
+      - name: Install Python dependencies and irrevolutions
+        run: |
+          echo "Installing Python dependencies and irrevolutions"
+          python3 -m pip install --upgrade pip
+          python3 -m pip install -r requirements.txt  # Adjust this if you're using a different file for dependencies
+          python3 -m pip install .  # Install irrevolutions package
+
+      # Optionally run tests within the environment
+      - name: Run tests
+        run: |
+          echo "Running tests"
+          conda activate fenicsx-env
+          cd test && pytest -v . || echo "Tests failed"
