### 2024-10-07 kumiori/irrevolutions#67 @kumiori
A license header has been added to the two files from dolfiny, with a pointer to the software package, the names of the authors, and the license text.
