### 2024-10-07 7cba41c
simplifying workflow
