### 2024-10-07 f577740
fixing conda relic
