+++ .github/workflows/ubuntu.yml
@@ -38,6 +38,7 @@ jobs:
         run: |
           echo "Installing Python dependencies and irrevolutions"
           python3 -m pip install --upgrade pip
+          python3 -m pip install --upgrade --force-reinstall numpy
           python3 -m pip install .  # Install irrevolutions package
 
       # Optionally run tests within the environment
