+++ .github/workflows/ubuntu.yml
@@ -38,7 +38,7 @@ jobs:
         run: |
           echo "Installing Python dependencies and irrevolutions"
           python3 -m pip install --upgrade pip
-          python3 -m pip install --upgrade --force-reinstall numpy
+          python3 -m pip install numpy==1.21.6 --force-reinstall # Downgrade NumPy to a compatible version
           python3 -m pip install .  # Install irrevolutions package
       - name: Check versions of installed packages
         run: |
