### 2024-10-07 e81f6b1
pinning numpy <2
