+++ pyproject.toml
@@ -18,6 +18,7 @@ dependencies = [
     "PyYAML>=6.0.1",
     "scipy>=1.12.0",
     "pyvista>=0.43",
+    "numpy==1.21.6",  # Pin NumPy to this version
     "numba>=0.60.0"
 ]
 readme = "README.md"
