### 2024-10-07 92fe8c6
numpy version
