### 2024-10-07 kumiori/irrevolutions#66 @kumiori
Following up on this issue, @jhale do you have any further comments on the paper?
The bold will be removed in the final version, it's used only to highlight the changes (are there other options/best practices?)
