### 2024-10-07 kumiori/irrevolutions#65 @kumiori
Thanks for the suggestion. The two Rayleigh scripts have been moved to the playground as indicated and the JOSS_paper_figures_rayleigh_benchmark notebook has been cleared from hardcoded signatures.

There is a step, however, where the user is requested to 'cherry pick' one of the output computations
via its hashed signature to produce a comparison figure, namely analytic vs. computation profiles.
This is done because some of the output profiles are 'trivial' (i.e., a constant) and may not be of interest
for the comparison.

Feel free to suggest how to improve the comparison, if you have further insights.
