### 2024-10-08 kumiori/irrevolutions#65 @jorgensd
Once this is added to a local readme file in `playground/rayleigh` as instructions, and you have decided on how to pick a non-trivial profile for visualization we are good here!:)
