### 2024-10-08 #6897 @jorgensd
Only very minor stuff left that @kumiori will fix very soon (on the reproducible figure part).
