+++ .github/workflows/workflow_dolfinx_container.yaml
@@ -9,7 +9,7 @@ on:
 jobs:
   set-up-computing-environment:
     runs-on: ubuntu-latest
-    container: ghcr.io/fenics/dolfinx/dolfinx:nightly
+    container: ghcr.io/fenics/dolfinx/lab:nightly
 
     steps:
       - name: Checkout repository
