### 2024-10-08 1580e32
reproducing JOSS figures, readme
