### 2024-10-08 kumiori/irrevolutions#65 @kumiori
Thanks for the snippet. The instructions have been added to a README.md file in the playground/rayleigh folder and the notebook file has been tweaked to automatically retrieve one nontrivial computation from the random batch.
This ensures that, following the provided instructions, a user is able to reproduce the paper's figures with a self-contained script.
