### 2024-10-08 #6897 @jorgensd
I guess once @jhale has reviewed the paper, changes, it should be change such that we do not see the diff with the old version
