### 2024-10-12 #6897 @majensen
If I understand you @jorgensd - if @kumiori makes the suggested changes and pushes them to the code repo, a new "generate pdf" should create a PDF that has those changes. Old PDFs in this thread won't be recreated. Is that the question?
