### 2024-10-12 #6897 @jorgensd
> If I understand you @jorgensd - if @kumiori makes the suggested changes and pushes them to the code repo, a new "generate pdf" should create a PDF that has those changes. Old PDFs in this thread won't be recreated. Is that the question?

Currently there is striked out text in the generated pdf, as @kumiori has done this to easily indicate the changes for the reviewers. I am very happy he did this, but as at the paper from my end is approved, this highlighting should be removed, such that the paper is ready for publishing. I would just like @jhale to indicate that he is fine with the changes before we proceed.
