### 2024-10-14 #6897 @majensen
@jorgensd - ah I see now. @jhale can you look at the changes in the PDF of https://github.com/openjournals/joss-reviews/issues/6897#issuecomment-2399757028, and comment here? If you approve, then @kumiori can proceed to clean up the paper for acceptance. Thank you!
