### 2024-10-29 #6897 @kumiori
Hello,
I am reaching out to inquire about the progress of our paper in review. Specifically, @jhale I wanted to check in regarding the two remaining open issues that were highlighted, on the licencing and the final `paper issues`.
