### 2024-11-16 acbd25f
removing bold and strikethrough for final acceptance
