### 2024-11-16 #6897 @kumiori
Dear Editor @majensen,
after clearing the issues and receiving the positive recommendations of @jhale and @jorgensd, the paper has been cleaned (from the diff in bold) and is ready for final decision.

Best
