### 2024-11-17 #6897 @majensen
Thanks @kumiori  - I will do a readthrough of the paper, this may generate minor suggestions. Can you please create an open archive of the repository using Zenodo, FigShare or similar? Here is [a short tutorial](https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729) using Zenodo. Please report the resulting DOI here. Thank you!
