### 2024-11-20 #6897 @majensen
Thanks @kumiori  - JOSS prefers the paper archive to have the same name as the paper. Can you update this?
