### 2024-11-21 #6897 @kumiori
Yes, @majensen I have updated the title of the package to match the JOSS submission.
The name of the compressed archive, however, is <name of the python package>-<version>.
Is this the correct way to go?
