### 2024-11-21 #6897 @majensen
This looks good to me @kumiori - the Title field is the location where JOSS needs the paper name.
