### 2024-11-23 #6897 @majensen
@kumiori - Thank you for this interesting contribution, which appears to me to solve a difficult phase space search problem numerically in order to guarantee finding single solutions for a wide class of systems.

I am sure the Associate Editor-in-Chief will think the paper as a whole is too long. Since JOSS focuses on the software, I would suggest putting the Verification and Appendix sections in a separate document in the code documentation repository, and only briefly describe their content in the paper itself.

The references in the text need some cleaning up. For example, in lines 20-21, rather than a format like "(Name, Year), ..., and (Name, Year)", we should see a format like "Name (Year), ... and Name (Year)."

The references for Kim Pham are inconsistent - in the text (e.g. line 21) we see "Kim Pham et al" and "K. Pham & Marigo", while in the Reference section we see (e.g.) "Pham, Kim,..." and "Pham, K....". Can you please rectify? Should these be "Pham et al", "Pham & Marigo"?

Here are some minor suggestions:

line 1: 𝜕’Alembert -> d’Alembert? universités -> université?
line 9: “We dedicate this contribution to releasing” -> “With this contribution, we release”
line 12: “patterns which is commonplace” -> “patterns. Systems with these qualities are commonplace”
line 24: “investigate up to second order” -> “investigate solutions up to second order” ?
line 84: “parameters)” - put closing paren on following line?
line 91:  industry standard is in italics - make roman
