+++ paper/paper.bib
@@ -104,7 +104,7 @@
 
 @article{Pham2013aa,
 	abstract = {Considering a family of gradient-enhanced damage models and taking advantage of its variational formulation, we study the stability of homogeneous states in a full three-dimensional context. We show that gradient terms have a stabilizing effect, but also how those terms induce structural effects. We emphasize the great importance of the type of boundary conditions, the size and the shape of the body on the stability properties of such states.},
-	author = {Pham, K. and Marigo, J. -J.},
+	author = {Pham, Kim and Marigo, Jean-Jacques},
 	date = {2013/01/01},
 	date-added = {2024-08-02 14:24:09 +0900},
 	doi = {10.1007/s10659-012-9382-5},
@@ -198,7 +198,7 @@
 
 @article{pham:2011-the-issues,
 	abstract = {We consider a wide class of gradient damage models which are characterized by two constitutive functions after a normalization of the scalar damage parameter. The evolution problem is formulated following a variational approach based on the principles of irreversibility, stability and energy balance. Applied to a monotonically increasing traction test of a one-dimensional bar, we consider the homogeneous response where both the strain and the damage fields are uniform in space. In the case of a softening behavior, we show that the homogeneous state of the bar at a given time is stable provided that the length of the bar is less than a state dependent critical value and unstable otherwise. However, we also show that bifurcations can appear even if the homogeneous state is stable. All these results are obtained in a closed form. Finally, we propose a practical method to identify the two constitutive functions. This method is based on the measure of the homogeneous response in a situation where this response is stable without possibility of bifurcation, and on a procedure which gives the opportunity to detect its loss of stability. All the theoretical analyses are illustrated by examples.},
-	author = {Kim Pham and Jean-Jacques Marigo and Corrado Maurini},
+	author = {Pham, Kim and Marigo, Jean-Jacques and Maurini,  Corrado},
 	date-added = {2024-07-31 14:57:29 +0900},
 	doi = {10.1016/j.jmps.2011.03.010},
 	issn = {0022-5096},
