+++ paper/addendum.md
@@ -1,0 +1,187 @@
+---
+title: Addedndum to 'Variational Solvers for Irreversible Evolutionary Systems'
+tags:
+    - Python
+    - evolutions
+    - stability
+    - bifurcation
+    - irreversibility
+    - nonlinear
+    - nonconvex
+    - singular perturbations
+authors:
+    - name: Andrés A {León Baldelli}
+      corresponding: true
+      orcid: 0000-0002-3019-602X
+      equal-contrib: false
+      affiliation: 1
+    - name: Pierluigi Cesana
+      orcid: 0000-0002-0304-9382
+      corresponding: true
+      equal-contrib: false
+      affiliation: 2
+affiliations:
+    - name: $\partial$'Alembert Institute, CNRS, Sorbonne Universités, Place Jussieu 75252 Paris Cedex 05, France
+      index: 1
+    - name: Institute of Mathematics for Industry, Kyushu University, 744 Motooka, Nishi-ku, Fukuoka, 819-0395, Japan.
+      index: 2
+date: 1 August 2024
+bibliography: paper.bib
+geometry: 'left=1.9cm,right=1.9cm,top=1.9cm,bottom=1.9cm'
+output: pdf_document
+---
+
+## Verification
+
+We benchmark our solvers against a nontrivial 1d problem (cf. `test/test_rayleigh.py` in the code repository), namely we compute
+
+$$
+\begin{aligned}
+&\min_{X_0} \mathcal R(z) \qquad \qquad [2.1] \\
+&\min_{\mathcal K^+_0} \mathcal R(z) \qquad\qquad [2.2]
+\end{aligned}
+$$
+
+where $z = (v, \beta)$ in $X_0$ and $\mathcal K^+_0$, using `BifurcationSolver` and `StabilitySolver`, respectively. The quantity $\mathcal R(z)$ is a Rayleigh ratio, often used in structural mechanics as a dimensionless global quantity (an energetic ratio of elastic and fracture energies) which provides insight into the stability and critical loading conditions for a structure. For definiteness, using the Sobolev spaces which are the natural setting for second order PDE problems, we set $X_0 = H^1_0(0, 1) \times H^1(0, 1)$ and $\mathcal K^+_0 = H^1_0(0, 1) \times \{\beta \in H^1(0, 1), \beta \geq 0\}$. Let
+$$\mathcal R(z):= \dfrac{\int_0^1 a(\beta'(x))^2dx+\int_0^1 b(v'(x) -c\beta(x))^2dx}{\int_0^1\beta(x)^2dx},\qquad\qquad [3]$$
+where $a, b, c$ are real coefficients such that $a>0, b>0, c\neq 0$. The quantity above occurs in the stability analysis of a 1d damageable bar, where $b$ is related to the spring constant of the material while $a, c$ encapsulate material, loading, and model parameters, cf. the Appendix of [@pham:2011-the-issues].
+
+\autoref{fig:profile}-\autoref{fig:phase_diag_cone} below provide a visual representation of the accuracy of the solvers in the solution of problems [2.1] and [2.2] by comparison to analytic solutions.
[…]
