### 2024-11-26 #6897 @jhale
Trying to be helpful, but with slight fear of overstepping, the Latin spelling I've used when submitting with your colleagues @kumiori is L'Institut Jean Le Rond d'Alembert. I feel like the Greek 𝜕 is more for the logo/graphic design.
