+++ playground/IDRIS-CAMPAING-HYDRA/scripts/1d-film-second-order-bifurcation-natural.py
@@ -7,9 +7,14 @@ import sys
 from pathlib import Path
 from typing import Optional
 
-from libidris.core import (elastic_energy_density_film, damage_energy_density, stress)
+from libidris.core import elastic_energy_density_film, damage_energy_density, stress
 from libidris.core import a
-from libidris.core import setup_output_directory, save_parameters, create_function_spaces_1d, initialize_functions
+from libidris.core import (
+    setup_output_directory,
+    save_parameters,
+    create_function_spaces_1d,
+    initialize_functions,
+)
 
 
 """
@@ -76,8 +81,15 @@ import pyvista
 import ufl
 import yaml
 from dolfinx.common import list_timings
-from dolfinx.fem import (Constant, Function, assemble_scalar, dirichletbc,
-                         form, locate_dofs_geometrical, set_bc)
+from dolfinx.fem import (
+    Constant,
+    Function,
+    assemble_scalar,
+    dirichletbc,
+    form,
+    locate_dofs_geometrical,
+    set_bc,
+)
 from dolfinx.fem.petsc import assemble_vector, set_bc
 from dolfinx.io import XDMFFile
 from irrevolutions.algorithms.am import HybridSolver
@@ -85,17 +97,25 @@ from irrevolutions.algorithms.so import BifurcationSolver, StabilitySolver
 
 from irrevolutions.solvers.function import vec_to_functions
 from irrevolutions.test.test_1d import _AlternateMinimisation1D as am1d
-from irrevolutions.utils import (ColorPrint, ResultsStorage, Visualization,
-                                 _logger, _write_history_data, history_data,
-                                 norm_H1, norm_L2)
-from irrevolutions.utils.plots import (plot_AMit_load, plot_energies,
-                                       plot_force_displacement)
-from irrevolutions.utils.viz import (plot_mesh, plot_profile, plot_scalar,
-                                     plot_vector)
+from irrevolutions.utils import (
+    ColorPrint,
+    ResultsStorage,
+    Visualization,
+    _logger,
+    _write_history_data,
+    history_data,
+    norm_H1,
+    norm_L2,
+)
+from irrevolutions.utils.plots import (
+    plot_AMit_load,
+    plot_energies,
+    plot_force_displacement,
+)
+from irrevolutions.utils.viz import plot_mesh, plot_profile, plot_scalar, plot_vector
 from mpi4py import MPI
 from petsc4py import PETSc
-from pyvista.utilities import xvfb
-
+from pyvista.plotting.utilities import xvfb
 from irrevolutions.utils.viz import _plot_bif_spectrum_profiles
 
 petsc4py.init(sys.argv)
@@ -117,8 +137,8 @@ def run_computation(parameters, storage=None):
     # N = max(_N, parameters["model"]["ell"] / parameters["geometry"]["mesh_size_factor"])
     N = parameters["model"]["ell"] / parameters["geometry"]["mesh_size_factor"]
     logging.info(f"Mesh size: {N}")
-    
-    mesh = dolfinx.mesh.create_unit_interval(MPI.COMM_WORLD, int(1/N))
+
+    mesh = dolfinx.mesh.create_unit_interval(MPI.COMM_WORLD, int(1 / N))
     outdir = os.path.join(os.path.dirname(__file__), "output")
 
     prefix = setup_output_directory(storage, parameters, outdir)
@@ -148,12 +168,11 @@ def run_computation(parameters, storage=None):
     # Define the state
     zero_u = Function(V_u, name="BoundaryUnknown")
     zero_u.interpolate(lambda x: np.zeros_like(x[0]))
-    
+
     u_zero = Function(V_u, name="InelasticDisplacement")
-    eps_t = dolfinx.fem.Constant(mesh, np.array(1., dtype=PETSc.ScalarType))
[…]
