+++ playground/pizza-notch/pizza-notch.py
@@ -15,23 +15,35 @@ import petsc4py
 import pyvista
 import ufl
 import yaml
-from dolfinx.fem import (Function, FunctionSpace, assemble_scalar, dirichletbc,
-                         form, locate_dofs_topological, set_bc)
+from dolfinx.fem import (
+    Function,
+    FunctionSpace,
+    assemble_scalar,
+    dirichletbc,
+    form,
+    locate_dofs_topological,
+    set_bc,
+)
 from dolfinx.io import XDMFFile, gmshio
 from dolfinx.mesh import locate_entities_boundary
 from mpi4py import MPI
 from petsc4py import PETSc
-from pyvista.utilities import xvfb
-
+from pyvista.plotting.utilities import xvfb
 from irrevolutions.algorithms.am import HybridSolver
 from irrevolutions.algorithms.so import BifurcationSolver, StabilitySolver
 from irrevolutions.meshes.pacman import mesh_pacman
 from irrevolutions.models import DamageElasticityModel as Brittle
-from irrevolutions.utils import (ColorPrint, _logger, _write_history_data,
-                                 history_data, set_vector_to_constant)
+from irrevolutions.utils import (
+    ColorPrint,
+    _logger,
+    _write_history_data,
+    history_data,
+    set_vector_to_constant,
+)
 from irrevolutions.utils.lib import _local_notch_asymptotic
 from irrevolutions.utils.viz import plot_mesh, plot_scalar, plot_vector
 import basix.ufl
+
 description = """We solve here a basic 2d of a notched specimen.
 Imagine a dinner a pizza which is missing a slice, and lots of hungry friends
 that pull from the sides of the pizza. Will a real pizza will break at the centre?
@@ -193,7 +205,6 @@ def run_computation(parameters, storage):
     stability = StabilitySolver(
         total_energy, state, bcs, cone_parameters=parameters.get("stability")
     )
-
 
     _logger.setLevel(level=logging.CRITICAL)
 
