### 2024-11-26 #6897 @kumiori
thank you @jhale I double-checked and changed the affiliation to the suggested
Sorbonne Université, CNRS, Institut Jean Le Rond d’Alembert, F-75005 Paris, France
the change has been pushed upstream and merged into main
