### 2024-11-27 #6897 @majensen
@kumiori - thanks for your work on this. I like the Addendum solution. I'm going to mention a few more suggestions/typos here, but I will make the recommendation.

line 70: "wrapping" -> "to wrap"
line 115: "1d" => "1D"
line 142: "Stmuctures" -> "structures"
line 142: De-capitalize journal article title? i.e., "Stable states and paths..." (consistent with journal article titles elsewhere in refs)
line 169: "Rate-independent systems" -> "Rate-independent Systems" (book title)
line 172: "Des séances de l'Académie Des" -> "des Séances de l'Académie des"
line 177-8: Capitalize words in book title? i.e. "Continuum Thermodynamics: The Art and Science..." (consistent with capitalization of journal titles elsewhere in refs)
line 182-3: De-capitalize journal article title? i.e. "Stability of homogeneous states..." (consistent with journal article titles elsewhere)
