### 2024-11-27 #6897 @majensen
@kumiori - can you also review the Missing and Invalid DOIs in https://github.com/openjournals/joss-reviews/issues/6897#issuecomment-2504881937?
Also, my guess is that [this bib line](https://github.com/kumiori/irrevolutions/blob/43e415849c9a3bf549f930b666ac0609a91ba8de/paper/paper.bib#L361) may be causing the [error above](https://github.com/openjournals/joss-reviews/issues/6897#issuecomment-2504883213).
