### 2024-11-29 b0df1d4
variable thickness decorator
