+++ src/irrevolutions/algorithms/am.py
@@ -11,8 +11,7 @@ from petsc4py import PETSc
 from irrevolutions.solvers import SNESSolver
 from irrevolutions.solvers.function import functions_to_vec
 from irrevolutions.solvers.snesblockproblem import SNESBlockProblem
-from irrevolutions.utils import (ColorPrint, norm_H1, norm_L2,
-                                 set_vector_to_constant)
+from irrevolutions.utils import ColorPrint, norm_H1, norm_L2, set_vector_to_constant
 from dolfinx.fem.petsc import assemble_vector, set_bc
 
 comm = MPI.COMM_WORLD
@@ -20,14 +19,15 @@ comm = MPI.COMM_WORLD
 # Set up basic logging
 logging.basicConfig()
 
+
 class AlternateMinimisation:
     """
     First order Alternate Minimisation solver for elasticity and damage fields.
 
-    The solver seeks equilibia (critical states for the total energy) of a system 
+    The solver seeks equilibia (critical states for the total energy) of a system
     by alternating between solving for the elasticity and the damage field in an iterative process,
     exploiting the separate convexity of the total energy.
-    
+
     Remark: check the assumptions.
 
     Parameters:
@@ -38,7 +38,7 @@ class AlternateMinimisation:
         Dictionary containing the current state of the system with keys 'u' (displacement)
         and 'alpha' (damage).
     bcs : dict
-        Dictionary of boundary conditions with keys 'bcs_u' for displacement and 'bcs_alpha' 
+        Dictionary of boundary conditions with keys 'bcs_u' for displacement and 'bcs_alpha'
         for damage.
     solver_parameters : dict, optional
         Dictionary of solver parameters, default is an empty dictionary.
@@ -109,16 +109,24 @@ class AlternateMinimisation:
         Parameters:
         ----------
         outdir : str, optional
-            Output directory for saving intermediate results. If provided, the mesh 
+            Output directory for saving intermediate results. If provided, the mesh
             and functions will be written to an XDMF file.
         """
 
         alpha_diff = dolfinx.fem.Function(self.alpha.function_space)
         self.data = {
-            "iteration": [], "error_alpha_L2": [], "error_alpha_H1": [], "F_norm": [],
-            "error_alpha_max": [], "error_residual_F": [], "error_residual_u": [],
-            "solver_alpha_reason": [], "solver_alpha_it": [], "solver_u_reason": [],
-            "solver_u_it": [], "total_energy": []
+            "iteration": [],
+            "error_alpha_L2": [],
+            "error_alpha_H1": [],
+            "F_norm": [],
+            "error_alpha_max": [],
+            "error_residual_F": [],
+            "error_residual_u": [],
+            "solver_alpha_reason": [],
+            "solver_alpha_it": [],
+            "solver_u_reason": [],
+            "solver_u_it": [],
+            "total_energy": [],
         }
 
         # Write the initial mesh
@@ -131,11 +139,13 @@ class AlternateMinimisation:
             ) as file:
                 file.write_mesh(self.u.function_space.mesh)
 
-        for iteration in range(1, self.solver_parameters.get("damage_elasticity").get("max_it")):
+        for iteration in range(
+            1, self.solver_parameters.get("damage_elasticity").get("max_it")
+        ):
             # Elasticity solver step
             with dolfinx.common.Timer("~First Order: AltMin-Elastic solver"):
                 (solver_u_it, solver_u_reason) = self.elasticity.solve()
-            
+
             # Damage solver step
[…]
