### 2024-12-03 #6897 @kumiori
thank you @majensen for providing your guess for the offending entry.

I have fixed missing doi's and the ISBN issue.

"Décomposition orthogonale d’un espace hilbertien ..." seems to lack a doi and I failed to associate a doi to "PETSc Web page" and "dolfiny: Python wrappers ..."
