### 2024-12-03 #6897 @majensen
Thanks @kumiori - the DOIs that the bot can't find shouldn't cause a problem. Will try again now!
