### 2024-12-03 #6897 @majensen
Hmm @kumiori - still seems to be an ISBN length problem. If you don't find anything upon reexamination, we can reach out to the JOSS developer for help.
