### 2024-12-04 e888af7
offending 9 char long isbn
