+++ paper/paper.bib
@@ -121,7 +121,6 @@
 	date-added = {2024-08-02 14:24:09 +0900},
 	doi = {10.1007/s10659-012-9382-5},
 	id = {Pham2013},
-	isbn = {1573-2681},
 	journal = {Journal of Elasticity},
 	number = {1},
 	pages = {63--93},
@@ -337,7 +336,6 @@
 	date-added = {2021-05-16 16:45:21 +0200},
 	date-modified = {2024-08-16 14:51:57 +0900},
 	doi = {10.1016/j.jmps.2021.104424},
-	isbn = {0022-5096},
 	journal = {Journal of the Mechanics and Physics of Solids},
 	keywords = {Damage; Fracture; Stability; Bifurcation; Variational methods},
 	pages = {104424},
@@ -381,7 +379,6 @@
 	date-modified = {2024-08-16 14:54:37 +0900},
 	doi = {10.1007/s10589-008-9167-8},
 	id = {Pinto da Costa2010},
-	isbn = {1573-2894},
 	journal = {Computational Optimization and Applications},
 	number = {1},
 	pages = {25--57},
