+++ paper/paper.bib
@@ -276,7 +276,6 @@
 	journal = {Comptes Rendus M{\'e}canique},
 	title = {La m{\'e}canique de l'endommagement au secours de la m{\'e}canique de la rupture : l'{\'e}volution de cette id{\'e}e en un demi-si{\`e}cle},
 	year = {2023},
-	bdsk-file-1 = {YnBsaXN0MDDSAQIDBFxyZWxhdGl2ZVBhdGhZYWxpYXNEYXRhbxBcAC4ALgAvAC4ALgAvAC4ALgAvAC4ALgAvAC4ALgAvAEwAaQBiAHIAYQByAHkALwBNAG8AYgBpAGwAZQAgAEQAbwBjAHUAbQBlAG4AdABzAC8AYwBvAG0AfgBhAHAAcABsAGUAfgBDAGwAbwB1AGQARABvAGMAcwAvAFAAYQBwAGUAcgBzAC8AbABlAG8DAQBuACAAYgBhAGwAZABlAGwAbABpAC0AMgAwADIAMQAgADIALgBwAGQAZk8RAcwAAAAAAcwAAgAADE1hY2ludG9zaCBIRAAAAAAAAAAAAAAAAAAAAOGUkbtCRAAB/////xhsZZduIGJhbGRlbGxpLTIwMjEgMi5wZGYAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAD/////3xvc9wAAAAAAAAAAAAUABQAACiBjdQAAAAAAAAAAAAAAAAAGUGFwZXJzAAIAXy86VXNlcnM6a3VtaW9yaTM6TGlicmFyeTpNb2JpbGUgRG9jdW1lbnRzOmNvbX5hcHBsZX5DbG91ZERvY3M6UGFwZXJzOmxlb8yBbiBiYWxkZWxsaS0yMDIxIDIucGRmAAAOADQAGQBsAGUAbwMBAG4AIABiAGEAbABkAGUAbABsAGkALQAyADAAMgAxACAAMgAuAHAAZABmAA8AGgAMAE0AYQBjAGkAbgB0AG8AcwBoACAASABEABIAXVVzZXJzL2t1bWlvcmkzL0xpYnJhcnkvTW9iaWxlIERvY3VtZW50cy9jb21+YXBwbGV+Q2xvdWREb2NzL1BhcGVycy9sZW/MgW4gYmFsZGVsbGktMjAyMSAyLnBkZgAAEwABLwAAFQACAA///wAAAAgADQAaACQA3wAAAAAAAAIBAAAAAAAAAAUAAAAAAAAAAAAAAAAAAAKv},
 	bdsk-url-1 = {https://www.sciencedirect.com/science/article/pii/S0022509621001010},
 	bdsk-url-2 = {https://doi.org/10.1016/j.jmps.2021.104424}
 	}
@@ -343,7 +342,6 @@
 	ty = {JOUR},
 	volume = {152},
 	year = {2021},
[…]
