### 2024-12-04 #6897 @kumiori
@majensen, I reexamined the bib file, it should be fixed now
