### 2024-12-04 d4ec87a
small changes in paper.bib
