+++ paper/paper.bib
@@ -26,8 +26,7 @@
 	author = {Zden\v{e}k P. Ba\v{z}ant},
 	doi = {10.1061/(ASCE)0733-9399(1988)114:12(2013)},
 	journal = {Journal of Engineering Mechanics},
-	title = {{Stable States and Paths of Stmuctures with Plasticity or Damage}
-	},
+	title = {Stable States and Paths of Stmuctures with Plasticity or Damage},
 	volume = {114},
 	year = {1988},
 	bdsk-url-1 = {https://www.sciencedirect.com/science/article/pii/S002250961100055X},
@@ -124,8 +123,7 @@
 	journal = {Journal of Elasticity},
 	number = {1},
 	pages = {63--93},
-	title = {{Stability of Homogeneous States with Gradient Damage Models: Size Effects and Shape Effects in the Three-Dimensional Setting}
-	},
+	title = {Stability of Homogeneous States with Gradient Damage Models: Size Effects and Shape Effects in the Three-Dimensional Setting},
 	volume = {110},
 	year = {2013},
 	bdsk-url-1 = {https://doi.org/10.1007/s10659-012-9382-5}
@@ -192,8 +190,7 @@
 	note = {New Computational Methods and Software Tools},
 	number = {9},
 	pages = {1124 - 1139},
-	title = {{Parallel distributed computing using Python}
-	},
+	title = {Parallel distributed computing using {P}ython},
 	volume = {34},
 	year = {2011},
 	bdsk-url-1 = {https://doi.org/10.1016/j.advwatres.2011.04.013}
@@ -379,8 +376,7 @@
 	journal = {Computational Optimization and Applications},
 	number = {1},
 	pages = {25--57},
-	title = {{Cone-constrained eigenvalue problems: theory and algorithms}
-	},
+	title = {Cone-constrained eigenvalue problems: theory and algorithms},
 	volume = {45},
 	year = {2010},
 	bdsk-url-1 = {https://doi.org/10.1007/s10589-008-9167-8}
