### 2024-12-04 #6897 @danielskatz
@kumiori - I've suggested changes for your paper in https://github.com/kumiori/irrevolutions/pull/102 - please merge this, or let me know what you disagree with, then we can continue the acceptance and publication process.
