### 2024-12-04 #6897 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
