+++ paper/paper.bib
@@ -101,18 +101,13 @@
 	year = {1994},
 	bdsk-url-1 = {https://doi.org/10.1115/1.3111068}
 	}
-
-@unpublished{camilla,
-	author = {Camilla Zolesi and Corrado Maurini},
-	date-added = {2024-08-02 16:57:06 +0900},
-	date-modified = {2024-08-02 16:58:32 +0900},
-	doi = {10.1016/j.jmps.2024.105802},
-	journal = {preprint, \url{https://hal.sorbonne-universite.fr/hal-04552309}
-	},
-	title = {Stability and crack nucleation in variational phase-field models of fracture: effects of length-scales and stress multi-axiality}
-	}
-	year = {}
-	}
+@article{zolesi:2024a,
+	author = {Zolesi, Camilla and Maurini, Corrado},
+	journal = {Journal of the Mechanics and Physics of Solids},
+	pages = {105802},
+	title = {Stability and crack nucleation in variational phase-field models of fracture: Effects of length-scales and stress multi-axiality},
+	volume = {192},
+	year = {2024}}
 
 @article{Pham2013aa,
 	abstract = {Considering a family of gradient-enhanced damage models and taking advantage of its variational formulation, we study the stability of homogeneous states in a full three-dimensional context. We show that gradient terms have a stabilizing effect, but also how those terms induce structural effects. We emphasize the great importance of the type of boundary conditions, the size and the shape of the body on the stability properties of such states.},
