### 2024-12-04 #6897 @danielskatz
👋 @kumiori - Please confirm that [the paper draft](https://raw.githubusercontent.com/openjournals/joss-papers/joss.06897/joss.06897/10.21105.joss.06897.pdf) looks correct to you now.
