### 2024-12-05 ee55b98
minor, dictionary representation
