### 2024-12-05 #6897 @kumiori
@danielskatz - I confirm that the draft looks good to the authors.
Best
