### 2024-12-05 d216050
Update CITATION.cff - published
