+++ CITATION.cff
@@ -13,7 +13,10 @@ authors:
     email: leon.baldelli@cnrs.fr
     affiliation: '∂’Alembert Institute, CNRS, Sorbonne Universités'
     orcid: 'https://orcid.org/0000-0002-3019-602X'
-repository-code: 'https://github.com/kumiori/mec647/'
+- family-names: Cesana
+  given-names: Pierluigi
+  orcid: "https://orcid.org/0000-0002-0304-9382"
+repository-code: 'https://github.com/kumiori/irrevolutions/'
 url: 'https://irreversible-variational-solvers.readthedocs.io/'
 abstract: >-
   To study irreversible evolutionary processes with a
@@ -35,3 +38,26 @@ keywords:
   - singular perturbations
   - fracture
 license: GPL-3.0
+doi: 10.5281/zenodo.14187154
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: León Baldelli
+    given-names: Andrés A
+    orcid: "https://orcid.org/0000-0002-3019-602X"
+  - family-names: Cesana
+    given-names: Pierluigi
+    orcid: "https://orcid.org/0000-0002-0304-9382"
+  date-published: 2024-12-05
+  doi: 10.21105/joss.06897
+  issn: 2475-9066
+  issue: 104
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 6897
+  title: Variational Solvers for Irreversible Evolutionary Systems
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.06897"
+  volume: 9
