### 2024-12-05 2d56e22
Update CITATION.cff
