### 2024-12-05 #6897 @danielskatz
Congratulations to @kumiori (Andrés A. León Baldelli) and co-authors on your publication!!

And thanks to @jorgensd, @jhale for reviewing, and to @majensen for editing!
JOSS depends on volunteers and we couldn't succeed with you
